% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-align_assign.R
\name{alignAssign}
\alias{alignAssign}
\title{Align a highlighted region's assignment operators.}
\usage{
alignAssign(rgx_op = NULL)
}
\arguments{
\item{rgx_op}{Regex for assignment operator}
}
\value{
Aligns the given or guessed operator within a highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
