% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{append_rownumber_col}
\alias{append_rownumber_col}
\title{Quickly append a column which contains the rownumber}
\usage{
append_rownumber_col(df)
}
\arguments{
\item{df}{tibble or data.frame}
}
\value{
df with additional column "rownumber"
}
\description{
Quickly append a column which contains the rownumber
}
