% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{appendObj}
\alias{appendObj}
\title{Append an objective function to a basic dMod.frame}
\usage{
appendObj(dMod.frame, prd = list(g * x * p), obj_data = list(normL2(data,
  prd, e)), obj = list(obj_data),
  pars = list(structure(rnorm(length(getParameters(obj))), names =
  getParameters(obj))), times = list(seq(min(as.data.frame(data)[["time"]]) *
  0.9, max(as.data.frame(data)[["time"]]) * 1.1, length.out = 100)), ...,
  keepCalls = F)
}
\arguments{
\item{keepCalls}{}
}
\description{
Append an objective function to a basic dMod.frame
}
\examples{
\dontrun{
appendObj(myframe) \%>\%
.[[".calls"]] \%>\%
  # str2 \%>\%
{.}
}
}
