% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{appendParframes}
\alias{appendParframes}
\title{Title}
\usage{
appendParframes(dMod.frame, parframes = list(as.parframe(fits)), ...,
  keepCalls = F)
}
\arguments{
\item{keepCalls}{}
}
\description{
Title
}
