% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{are_names_of}
\alias{are_names_of}
\title{Pipe-friendly assigning of vectors when their names are known first}
\usage{
are_names_of(char_vec, value)
}
\arguments{
\item{value}{}
}
\description{
Pipe-friendly assigning of vectors when their names are known first
}
