% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{as.condition.grid}
\alias{as.condition.grid}
\title{Set rownames so that the covtable is in condition.grid format}
\usage{
as.condition.grid(df)
}
\arguments{
\item{df}{a data.frame containing the covariates}
}
\description{
Set rownames so that the covtable is in condition.grid format
}
