% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{as.datalist.NULL}
\alias{as.datalist.NULL}
\title{As.datalist method for NULL}
\usage{
\method{as.datalist}{`NULL`}(...)
}
\arguments{
\item{...}{NULL}
}
\description{
As.datalist method for NULL
}
\details{
Not sure if I still need it, it was mainly for the convenient use in methatcetin_fitting.
}
