% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_parameterHandling.R
\name{build_parameters_df}
\alias{build_parameters_df}
\title{Template to build a basic parameters_df}
\usage{
build_parameters_df(odes, observables, errormodel)
}
\arguments{
\item{odes, observables, errormodel}{symbolic definitions of functions}
}
\value{
data.frame with columns name0,name = name0, value = 1, unit = "a.u.", upper = 1e3 ,lower = 1e-5, 
estscale = "L", initpar = F, dynpar = F, obspar = F, errpar = F
}
\description{
Template to build a basic parameters_df
}
