% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cf_apply_expression}
\alias{cf_apply_expression}
\title{magrittr-Style Apply an expression that's more complicated}
\usage{
cf_apply_expression(x, ...)
}
\arguments{
\item{expr}{}
}
\description{
Old, I use {.} now...
}
