% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cf_applyPaginate}
\alias{cf_applyPaginate}
\title{Get list of paginated plots}
\usage{
cf_applyPaginate(pl, paginateInfo)
}
\arguments{
\item{pl}{(not yet facetted) ggplot}

\item{paginateInfo}{output from \link{cf_paginateInfo}}
}
\value{
list of ggplots
}
\description{
use ggforce
}
\examples{
# adapted from ggforce examples
paginateInfo <- cf_paginateInfo(~cut:clarity)
plotlist <- cf_applyPaginate(pl, paginateInfo)
plotlist[[1]]
plotlist[[5]]
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
