% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_parameterHandling.R
\name{cf_build_pargrids}
\alias{cf_build_pargrids}
\title{Build basic fixed.grid, est.grid, est.vec_df}
\usage{
cf_build_pargrids(
  parameters_df,
  parameters_estimate,
  condition.grid,
  condition.grid_parcolumns = NULL
)
}
\arguments{
\item{parameters_df}{}

\item{parameters_estimate}{}

\item{condition.grid}{}

\item{condition.grid_parcolumns}{names of cols which have parameter values in them}
}
\value{

}
\description{
Build basic fixed.grid, est.grid, est.vec_df
}
