% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_parameterHandling.R
\name{cf_condition_specific_wrapper}
\alias{cf_condition_specific_wrapper}
\title{Title}
\usage{
cf_condition_specific_wrapper(condition_specific, est.grid, est.vec_df,
  condition.grid)
}
\arguments{
\item{condition_specific}{tibble with columns 
char "parname", 
list of chars "conditions" 
char "condition_column" => refers to colname in condition.grid
lgl FLAGdummifyOtherConds}

\item{est.grid}{}

\item{est.vec_df}{}

\item{condition.grid}{}
}
\description{
Title
}
