% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{cf_copy_script}
\alias{cf_copy_script}
\title{Rename a script, change filename occurences within script, rename output folder}
\usage{
cf_copy_script(
  from,
  to,
  FLAGremoveOld = FALSE,
  FLAGrenameOutputFolder = FLAGremoveOld
)
}
\arguments{
\item{from, to}{as in file.rename}
}
\description{
Rename a script, change filename occurences within script, rename output folder
}
