% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{cf_copy_scripts_multiple}
\alias{cf_copy_scripts_multiple}
\title{cf_copy_script for multiple scripts at once}
\usage{
cf_copy_scripts_multiple(filenames, FLAGrename = FALSE)
}
\arguments{
\item{filenames}{data.table(from,to)}
}
\value{
nothing, called for side-effect
}
\description{
cf_copy_script for multiple scripts at once
}
\examples{
data.table(`~from` = list.files(".", "S501"),`~to` = list.files(".", "S501")) \%>\% cfoutput_MdTable(NFLAGtribble = 2)
filenames <- data.table(tibble::tribble(
~from, ~to,
"S520-MLPSR-01-DataPreparation.R"     ,"S530-MLCLS-01-DataPreparation.R"     ,
"S520-MLPSR-02-ModelAgnostic.R"       ,"S530-MLCLS-02-ModelAgnostic.R"       ,
"S520-MLPSR-03-FitDataSetVariations.R","S530-MLCLS-03-FitDataSetVariations.R",
"S520-MLPSR-04-CollectModelEnsemble.R","S530-MLCLS-04-CollectModelEnsemble.R",
"S520-MLPSR-06-ExportTopPredictors.R" ,"S530-MLCLS-06-ExportTopPredictors.R" ,
))
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
