% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_indiv.R
\name{cf_datapointL2}
\alias{cf_datapointL2}
\title{DatapointL2 without the env bullshit}
\usage{
cf_datapointL2(
  name,
  time,
  value,
  sigma = 1,
  attr.name = "validation",
  condition,
  prd_indiv
)
}
\arguments{
\item{name}{}

\item{time}{}

\item{value}{}

\item{sigma}{}

\item{attr.name}{}

\item{condition}{}

\item{prd_indiv}{}
}
\value{

}
\description{
DatapointL2 without the env bullshit
}
