% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cf_evaluate_formals}
\alias{cf_evaluate_formals}
\title{Get the default values of formals into the Global environment}
\usage{
cf_evaluate_formals(..., indices = 1:length(formals(...)))
}
\arguments{
\item{...}{the function}

\item{indices}{the indices of the formals you want to evaluate}
}
\description{
Get the default values of formals into the Global environment
}
\details{
Useful for debugging
}
\examples{
evaluate_formals(lm, indices = 1:5)
}
