% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cf_global_env_without}
\alias{cf_global_env_without}
\title{List the elements of .GlobalEnv without elements with matching naes}
\usage{
cf_global_env_without(reg)
}
\arguments{
\item{reg}{A vector of regex to be matched}
}
\description{
List the elements of .GlobalEnv without elements with matching naes
}
