% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-full_setup.R
\name{cf_install_rstudio}
\alias{cf_install_rstudio}
\title{Quick setup of RStudio on a new machine}
\usage{
cf_install_rstudio(FLAGoverwrite = TRUE, FLAGshortcuts = FALSE,
  FLAGreservedWords = FALSE)
}
\value{
called for side-effect
}
\description{
Quick setup of RStudio on a new machine
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
