% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cf_is_in_between}
\alias{cf_is_in_between}
\title{magrittr-style two sided comparison}
\usage{
cf_is_in_between(x, lower_lim, upper_lim)
}
\arguments{
\item{upper_lim}{}
}
\description{
magrittr-style two sided comparison
}
