% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cf_normL2}
\alias{cf_normL2}
\title{normL2 with NULL allowed for data or x}
\usage{
cf_normL2(data, x, ...)
}
\arguments{
\item{...}{}
}
\description{
normL2 with NULL allowed for data or x
}
\examples{
obj_apap <- cf_normL2(data_apap \%>\% as.datalist, (g_apap*x_apap*p_apap))
obj_apap <- cf_normL2(NULL, (g_apap*x_apap*p_apap))
obj_apap <- cf_normL2(data_apap \%>\% as.datalist, NULL)
obj_apap(pars_apap)

# why is this?
NULL + obj_apap
obj_apap + NULL
}
