% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_indiv.R
\name{cf_normL2_indiv}
\alias{cf_normL2_indiv}
\title{Fast normL2}
\usage{
cf_normL2_indiv(data, prd0, errmodel = NULL, est.grid, fixed.grid,
  times = NULL, attr.name = "data", FLAGverbose = FALSE,
  FLAGbrowser = FALSE)
}
\arguments{
\item{data}{}

\item{prd0}{}

\item{errmodel}{}

\item{est.grid}{}

\item{fixed.grid}{}

\item{times}{}

\item{attr.name}{}
}
\value{
objective function
}
\description{
Fast normL2
}
