% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cf_paginateInfo}
\alias{cf_paginateInfo}
\title{Collector function for arguments to call paginate}
\usage{
cf_paginateInfo(
  facets,
  nrow = 2,
  ncol = 3,
  scales = "fixed",
  type = c("wrap", "grid"),
  ...
)
}
\arguments{
\item{facets}{}

\item{nrow}{}

\item{ncol}{}

\item{scales}{}

\item{type}{}

\item{...}{see \code{\link[ggforce:facet_wrap_paginate]{ggforce::facet_wrap_paginate()}} and \code{\link[ggforce:facet_grid_paginate]{ggforce::facet_grid_paginate()}}}
}
\value{

}
\description{
Collector function for arguments to call paginate
}
\examples{
# see ?cf_applyPaginate
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
