% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_parameterHandling.R
\name{cf_parameters_df_condition_specific}
\alias{cf_parameters_df_condition_specific}
\title{Make parameters_df parameters condition specific}
\usage{
cf_parameters_df_condition_specific(
  parameters_df,
  conditions = c("C1", "C2"),
  parnames = parameters_df$name
)
}
\arguments{
\item{parameters_df}{a typical parameters_df}

\item{conditions}{character vector.}

\item{parnames}{names of parameters to apply condition specificity to}
}
\value{
new paraemters_df with new rows
}
\description{
will be done like this: parname_condition
}
