% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_parameterHandling.R
\name{cf_parameters_df_merge_values}
\alias{cf_parameters_df_merge_values}
\title{Merge values into a parameters_df}
\usage{
cf_parameters_df_merge_values(pars_df_into, from, scale = c("outer",
  "inner"))
}
\arguments{
\item{pars_df_into}{parameters_df}

\item{from}{Either named vector or parameters_df}
}
\value{
pars_df_into with updated values
}
\description{
Merge values into a parameters_df
}
