% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cf_plotData.datalist}
\alias{cf_plotData.datalist}
\title{Copied from dMod with additional title showing covariates}
\usage{
cf_plotData.datalist(
  data,
  ...,
  scales = "free",
  facet = "wrap",
  transform = NULL,
  aesthetics = NULL
)
}
\arguments{
\item{data, ..., scales, facet, transform}{see dMod::plotData}
}
\description{
Copied from dMod with additional title showing covariates
}
