% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_indiv.R
\name{cf_PRD_indiv}
\alias{cf_PRD_indiv}
\title{Title}
\usage{
cf_PRD_indiv(prd0, est.grid, fixed.grid)
}
\arguments{
\item{prd0}{}

\item{est.grid}{}

\item{fixed.grid}{}
}
\value{

}
\description{
Title
}
