% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_indiv.R
\name{cf_predict}
\alias{cf_predict}
\title{predict.prdfn with options bowser and verbose}
\usage{
cf_predict(
  prd,
  times,
  pars,
  keep_names = NULL,
  ncores = 4,
  FLAGverbose = FALSE,
  FLAGverbose2 = FALSE,
  FLAGbrowser = FALSE,
  ...
)
}
\arguments{
\item{prd}{}

\item{pars}{parframe}

\item{keep_names}{}

\item{FLAGverbose2}{}
}
\value{
data.table
}
\description{
predict.prdfn with options bowser and verbose
}
