% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cf_print_r.named_vector}
\alias{cf_print_r.named_vector}
\title{nicely formatted dput for named vectors}
\usage{
cf_print_r.named_vector(myvec)
}
\arguments{
\item{myvec}{}
}
\description{
nicely formatted dput for named vectors
}
