% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cf_printpars}
\alias{cf_printpars}
\title{Print pars so that you can copy paste it in your script and directly modify them}
\usage{
cf_printpars(pars)
}
\arguments{
\item{pars}{}
}
\description{
Print pars so that you can copy paste it in your script and directly modify them
}
