% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_wrappers.R
\name{cf_profile}
\alias{cf_profile}
\title{Better wrapper around dMod::profile}
\usage{
cf_profile(
  obj,
  pars,
  whichPar,
  alpha = 0.05,
  limits = c(lower = -Inf, upper = Inf),
  method = c("optimize"),
  stepControl = list(limit = 40, min = log10(1.005), stepsize = log10(1.005)),
  algoControl = NULL,
  optControl = NULL,
  verbose = FALSE,
  cores = 1,
  path = tempdir(),
  FLAGoverwrite = FALSE,
  ...
)
}
\arguments{
\item{outputFolder}{}
}
\value{
list of profiles
}
\description{
Improvements
* Automatically save profiles with consistent filenames
* Run each profile in a try
}
\details{
inheritParams dMod::profile
}
