% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cf_profile_plotPathsAffected}
\alias{cf_profile_plotPathsAffected}
\title{Title}
\usage{
cf_profile_plotPathsAffected(
  profiles,
  tol = 1e-05,
  ggCallback = list(facet_wrap_paginate(~PARAMETER1 + PARAMETER2, nrow = 3, ncol = 4,
    scales = "free", page = 1), guides(color = TRUE)),
  filename = NULL,
  FLAGfuture = TRUE,
  width = 29.7,
  height = 21,
  scale = 1,
  units = "cm"
)
}
\arguments{
\item{profiles}{profiles to plotPaths}

\item{tol}{}

\item{page}{pagination page}
}
\value{
paginated ggplot
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
