% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cf_profile_prepareAffectedPaths}
\alias{cf_profile_prepareAffectedPaths}
\title{Prepare parameters affected by a profile}
\usage{
cf_profile_prepareAffectedPaths(profiles, tol = 1e-05)
}
\arguments{
\item{tol}{tolerance to measure if a parameter is affected}

\item{dp}{data.table with profiles}

\item{parnames}{columns indicating parameter names}
}
\value{
data.table(value, constraint, PARAMETER1, PARAMETER2, PARVALUE1, PARVALUE2)]
}
\description{
Its not helpful to plot all paths in plotPaths,
but only of those parameters which are affected
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
