% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{cf_script_identifyStale}
\alias{cf_script_identifyStale}
\title{Title}
\usage{
cf_script_identifyStale()
}
\value{
list of stale scripts and stale folders
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
