% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{cf_script_makeFilenames}
\alias{cf_script_makeFilenames}
\title{Generate filenames}
\usage{
cf_script_makeFilenames(pattern_from, pattern_to)
}
\arguments{
\item{pattern_from}{}

\item{pattern_to}{}
}
\value{
data.table to use in cf_copy_script_multiple
}
\description{
Generate filenames
}
\examples{
pattern_from <- "S240-0(\\\\d)-Viktor-Smads"
pattern_to <- "S242-0\\\\1-SmadsRec-linearTot"
}
