% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{cf_script_remove}
\alias{cf_script_remove}
\title{Remove scripts AND their output folders}
\usage{
cf_script_remove(filenames_vector)
}
\arguments{
\item{filenames_vector}{}
}
\value{

}
\description{
Remove scripts AND their output folders
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
