% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{cf_scripts_identifyStale}
\alias{cf_scripts_identifyStale}
\title{Title}
\usage{
cf_scripts_identifyStale()
}
\value{
list of stale scripts and stale folders
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
