% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{cf_slide_section}
\alias{cf_slide_section}
\title{Title}
\usage{
cf_slide_section(.slides, title, footer = "", slide_num = "")
}
\arguments{
\item{.slides}{}

\item{title}{}

\item{footer}{}

\item{slide_num}{}
}
\description{
Title
}
