% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{cf_slide_titlecontent}
\alias{cf_slide_titlecontent}
\title{Title}
\usage{
cf_slide_titlecontent(.slides, title, content, footer = "", slide_num = "")
}
\arguments{
\item{.slides}{}

\item{title}{}

\item{content}{}

\item{footer}{}

\item{slide_num}{}
}
\value{

}
\description{
Title
}
