% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{cf_slide_twocontent}
\alias{cf_slide_twocontent}
\title{Title}
\usage{
cf_slide_twocontent(
  .slides,
  title,
  content1,
  content2,
  footer = "",
  slide_num = ""
)
}
\arguments{
\item{.slides}{}

\item{title}{}

\item{content1}{}

\item{content2}{}

\item{footer}{}

\item{slide_num}{}
}
\value{

}
\description{
Title
}
