% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.R
\name{cf_slides}
\alias{cf_slides}
\title{Title}
\usage{
cf_slides(template = 1,
  templatelist = system.file("templates/Uni_Praesentation_E1_RGB_16_9.potx",
  package = "conveniencefunctions"))
}
\arguments{
\item{template}{}
}
\description{
Title
}
