% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cf_startDev}
\alias{cf_startDev}
\alias{cf_stopDev}
\title{Title}
\usage{
cf_startDev(filename, width = 14, height = 10, scale = 0.6, ...)

cf_stopDev(filename)
}
\arguments{
\item{filename}{}

\item{width, height, scale}{In inches}

\item{...}{Arguments going to either \code{\link[grDevices:pdf]{grDevices::pdf()}} or \code{\link[grDevices:png]{grDevices::png()}}}
}
\value{
nothing, called for side effect
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
