% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cf_str_subset_name}
\alias{cf_str_subset_name}
\title{subset a vector by matching names against a pattern}
\usage{
cf_str_subset_name(vec, pattern)
}
\arguments{
\item{pattern}{}
}
\description{
This function is inspired by str_subset from stringr
}
