% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{cfcolors}
\alias{cfcolors}
\title{Colors}
\format{
An object of class \code{character} of length 120.
}
\usage{
cfcolors
}
\description{
Colors
}
\examples{
ggplot(data.frame(x = 1:30,color = factor(cfcolors[1:30], unique(cfcolors[1:30])))) + 
  geom_tile(aes(x=x, y = 1, fill = color)) + 
  scale_color_manual(values= setNames(nm = cfcolors),
  aesthetics = c("color", "fill"))
}
\keyword{datasets}
