% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cfgg_getAllAesthetics}
\alias{cfgg_getAllAesthetics}
\title{Get comprehensive list of all aesthetics accepted by each geom}
\usage{
cfgg_getAllAesthetics()
}
\value{
list(geom_abline = c("slope", "intercept", ...), ...)
}
\description{
From Moody_Mudskipper at
https://stackoverflow.com/questions/11657380/is-there-a-table-or-catalog-of-aesthetics-for-ggplot2
}
\details{
License is "CC BY-SA 4.0" according to https://stackoverflow.com/help/licensing
}
