% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cfggplot}
\alias{cfggplot}
\title{cfggplot}
\usage{
cfggplot(data = NULL, mapping = aes())
}
\arguments{
\item{data, mapping}{see ?ggplot}
}
\value{
ggplot
}
\description{
cfggplot
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
