% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cfn}
\alias{cfn}
\title{*.fn in functional form for safe evaluation and allowing for p1 or p2 to be NULL}
\usage{
cfn(p1, p2)
}
\arguments{
\item{p2}{}
}
\description{
*.fn in functional form for safe evaluation and allowing for p1 or p2 to be NULL
}
