% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check}
\alias{check}
\title{Run expression and print OK/FAILED if expression returns TRUE/FALSE}
\usage{
check(message, expr, nchar = 80, FLAGthrowError = FALSE)
}
\arguments{
\item{message}{character, e.g., "if all subjects are in data".}

\item{expr}{the expression to be evaluated}

\item{nchar}{integer, maximum number of character for message (for nicer print-out)}
}
\description{
Daniel Kaschek's check function
}
