% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check2sink}
\alias{check2sink}
\title{Write the output of the check to a file}
\usage{
check2sink(path = "check.txt")
}
\arguments{
\item{path}{where the output should be sunk to}
}
\description{
Write the output of the check to a file
}
