% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{checkin_hypothesis}
\alias{checkin_hypothesis}
\title{Get elements from the .GlobalEnv and turn it into a row of the dMod.frame}
\usage{
checkin_hypothesis(dMod.frame, exclude = NULL, include = NULL,
  prefix = "", suffix = "")
}
\arguments{
\item{dMod.frame}{the dMod.frame}

\item{exclude}{not yet implemented}

\item{include}{}
}
\value{
the dMod.frame augmented by the new row
}
\description{
Get elements from the .GlobalEnv and turn it into a row of the dMod.frame
}
\examples{

remove("doedel")
dMod.frame <- tibble(doedel = "blabal")
checkin_hypothesis(dMod.frame)

doedel <- "yay"
checkin_hypothesis(dMod.frame)

doedel <- function(x) x^2
checkin_hypothesis(dMod.frame)

}
