% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{checkout_hypothesis}
\alias{checkout_hypothesis}
\title{dMod.frame -----}
\usage{
checkout_hypothesis(dMod.frame, hypothesis, prefix = "", suffix = "")
}
\arguments{
\item{hypothesis}{character or numeric. specifying the name  or the index of the hypothesis}
}
\description{
Example of a dMod.frame pipe
}
\details{
dMod.frame <- tibble(hypothesis = c("cell specific pars with prior", "no cell specifc pars, no prior"),
x = list(x,x),
g = list(g,g),
p = list(p_1, p_2),
pars = list(pars_1, pars2),
data = list(mydatalist, mydatalist)) %>%
  mutate(prd = list((g*x*p_1), (g*x*p_2))) %>%
  mutate(prior = list(constraintL2(mu1, sigma1), NULL)) %>%
  mutate(obj = lapply(seq_along(x), function(i) normL2(data[[i]], prd[[i]]) + prior[[i]])) %>%
  mutate(machine = list(c("knecht1", "knecht2"), c("knecht3"))) %>%
  mutate(fits = lapply(seq_along(x), function(i) runbg(mstrust(obj[[i]], pars[[i]]), machine = machine[[i]]) ))
# wait a little
dMod.frame <- dMod.frame %>%
  mutate(fits = lapply(seq_along(x), function(i) fits[[i]][["get"]]()[[machine[[i]]]] )) %>%
  mutate(fits = lapply(fits, as.parframe)) %>%
  mutate(plots = lapply(seq_along(x), function(i) plotCombined(prd(mytimes, as.parvec(fits[[i]]), data = data[[i]])) ))

Load one row of a dMod.frame into the .GlobalEnv
}
\examples{
a <- 3
myfun <- function(x) {a * x^2}
testframe <- tibble(hypothesis = c("apap", "penner"),
                    schwurbel = list(qplot(1:10,1:10), "Penis"),
                    wupwup = list("moep", faithful),
                    myfun = list(function(x) {a * x^2}, myfun),
                    a = c(1:2))

checkout_hypothesis(testframe, "penner")
myfun
schwurbel
wupwup
hypothesis
}
