% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{checkSensitivities}
\alias{checkSensitivities}
\title{Check function to check sensitivities}
\usage{
checkSensitivities(prd, times, pars, whichpar, cond = 1, step = 0.1)
}
\arguments{
\item{prd}{prediction function}

\item{times}{vector of times}

\item{pars}{vector of pars}

\item{whichpar}{names of pars to check}

\item{cond}{indexing for condition}

\item{step}{stepsize for finite difference}
}
\description{
Check function to check sensitivities
}
\details{
Taken (with slight changes) from Daniel Kaschek in dMod/inst/examples/events.R
}
\seealso{
[do_checkSensitivities()]
}
\concept{analysisFunctions}
