% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{compare_named_numeric}
\alias{compare_named_numeric}
\title{compare named numerics}
\usage{
compare_named_numeric(.x, .y)
}
\arguments{
\item{.x, .y}{Named numerics}
}
\value{
data.frame with name, diff = (.y-.x), appended output of compare_names
}
\description{
compare named numerics
}
