% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{compare_names}
\alias{compare_names}
\title{Setdiff of names}
\usage{
compare_names(.x, .y)
}
\arguments{
\item{.x, .y}{Vectors/matrices with (dim)names}
}
\value{
output of compare.character
}
\description{
Setdiff of names
}
