% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{compare_objlist}
\alias{compare_objlist}
\title{compare two objlists}
\usage{
compare_objlist(.x, .y)
}
\arguments{
\item{.x, .y}{objlist}
}
\value{
difference of their values
}
\description{
compare two objlists
}
