% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{compare_vectors}
\alias{compare_vectors}
\title{compare two vectors}
\usage{
compare_vectors(x, y)
}
\arguments{
\item{x, y}{vectors to be compared}
}
\value{
only for printing
}
\description{
prints some stuff nicely
}
\examples{
compare_vectors(letters[1:10], letters[15:5])
}
