% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{construct_parframe}
\alias{construct_parframe}
\title{Helper function to reproducibly construct parframes}
\usage{
construct_parframe(pars, n = 20, seed = 12345, samplefun = rnorm)
}
\arguments{
\item{pars}{named vector. values don't play a role, only names}

\item{n}{how many lines should the parframe have}

\item{seed}{seed for the random number generator}

\item{samplefun}{rnorm, runif, etc...}
}
\value{
parframe (without metanames)
}
\description{
Helper function to reproducibly construct parframes
}
\examples{
construct_parframe(c(a = 0, b = 100000), 5)
construct_parframe(c(a = 0, b = 100000), 5)
}
