% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{copyFilesScript}
\alias{copyFilesScript}
\title{Create a script to copy files}
\usage{
copyFilesScript()
}
\value{
Cats the script to the console
}
\description{
Create a script to copy files
}
\examples{
copyFilesScript()
}
