% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cor2cov}
\alias{cor2cov}
\title{Turn correlation matrix and standard deviations back to covariance matrix}
\usage{
cor2cov(mycor, mysd = rep(1, dim(mycor)[1]))
}
\arguments{
\item{mycor}{correlation matrix}

\item{mysd}{standard deviations}
}
\value{
covariance matrix
}
\description{
Turn correlation matrix and standard deviations back to covariance matrix
}
\examples{
mycor <- matrix(c(1,0.8,0.8,1),2)
mysd <- c(sqrt(2),1)
cor2cov(mycor,mysd)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
