% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{d2d2dMod_data}
\alias{d2d2dMod_data}
\title{Bring wide d2d format of data into long dMod format}
\usage{
d2d2dMod_data(data, keep = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{keep}{character vector of columns with covariates to preserve and not count as variables}
}
\value{
The data in long format
}
\description{
The challenge is that sigmas get their own column each, therefore it's a bit more complicated than simply gathering
}
