% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial2.R
\name{d2d2r}
\alias{d2d2r}
\title{Import a model from the D2D definition files to dMod}
\usage{
d2d2r(setup_file = NULL, model_def_file = NULL, data_files = NULL,
  readme_file = NULL)
}
\arguments{
\item{setup_file}{A script which loads the model and data}

\item{model_def_file}{filename}

\item{data_files}{Character vector with file paths}

\item{readme_file}{filename}
}
\description{
Import a model from the D2D definition files to dMod
}
