% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{datatimes}
\alias{datatimes}
\title{create vector of times from min to max}
\usage{
datatimes(data, n = 500)
}
\arguments{
\item{data}{datalist}
}
\description{
create vector of times from min to max
}
