% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{describe_plotValue}
\alias{describe_plotValue}
\alias{describe_plotCombined}
\alias{describe_plotPars}
\alias{describe_plotPars}
\alias{describe_plotProfile}
\title{Describe plot}
\usage{
describe_plotValue(dMod.frame = NULL)

describe_plotCombined(dMod.frame = NULL)

describe_plotPars(dMod.frame = NULL)

describe_plotPars(dMod.frame = NULL)

describe_plotProfile(dMod.frame = NULL)
}
\arguments{
\item{dMod.frame}{}

\item{dMod.frame}{}
}
\description{
Here are some templates to motivate you to describe the plots.

Addin
}
\seealso{
Other insertfunctions: \code{\link{insert_seq_along_x_function_i}},
  \code{\link{insert_tee_operator}},
  \code{\link{insert_tee_print}},
  \code{\link{print_r.eqnlist2addReaction_pipe}},
  \code{\link{split_chunks}}

Other insertfunctions: \code{\link{insert_seq_along_x_function_i}},
  \code{\link{insert_tee_operator}},
  \code{\link{insert_tee_print}},
  \code{\link{print_r.eqnlist2addReaction_pipe}},
  \code{\link{split_chunks}}
}
