% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{dMf_append_plots}
\alias{dMf_append_plots}
\title{Add plots to the dMod.frame}
\usage{
dMf_append_plots(mydMod.frame, tol = 1)
}
\arguments{
\item{dMod.frame}{with cols x, parframes, data, best_prediction, hypothesis, best_value, conditions (optional)}
}
\value{
dMod.frame with plots plot_value, plot_pars, plot_combined
}
\description{
Add plots to the dMod.frame
}
