% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{dMf_expand_fits}
\alias{dMf_expand_fits}
\title{Expand a dMod.frame with parlist by columns derived from these fits}
\usage{
dMf_expand_fits(mydMod.frame)
}
\arguments{
\item{dMod.frame}{with columns x, data (list of datalists), prd}
}
\value{
dMod.frame augmented by columns parframes, best_parvec_best_value, best_prediction
}
\description{

}
