% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{dMod.frame}
\alias{dMod.frame}
\title{Generate a dMod.frame object}
\usage{
dMod.frame(hypothesis, g, x, p, data, e = NULL, ...)
}
\arguments{
\item{hypothesis}{Character. Description of the hypothesis}

\item{g}{fn}

\item{x}{fn}

\item{p}{fn}

\item{data}{data.frame or datalist, will be coerced to datalist}

\item{e}{fn}

\item{...}{}
}
\value{
Object of class \code{tbl_df}.
}
\description{
The dMod.frame object stores all objects that are needed to reproducibly
produce a result, such as a plot or profiles, in one row of a \link{tibble}.
Each row corresponds to a distinct hypothesis, e.g. one would have two distinct rows
for fitting with and without a prior.

Since the dMod.frame is also designed for scripting, the class will not be called
"dMod.frame" as I initially planned, but will be c("tbl_df", "tbl", "data.frame").
This way, I don't have to copy all the dplyr-verbs.
}
