% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\docType{data}
\name{dMod.frame0}
\alias{dMod.frame0}
\title{dMod.frame -----}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 10 columns.}
\usage{
dMod.frame0
}
\description{
Example of a dMod.frame pipe
}
\details{
dMod.frame <- tibble(hypothesis = c("cell specific pars with prior", "no cell specifc pars, no prior"),
x = list(x,x),
g = list(g,g),
p = list(p_1, p_2),
pars = list(pars_1, pars2),
data = list(mydatalist, mydatalist)) %>%
  mutate(prd = list((g*x*p_1), (g*x*p_2))) %>%
  mutate(prior = list(constraintL2(mu1, sigma1), NULL)) %>%
  mutate(obj = lapply(seq_along(x), function(i) normL2(data[[i]], prd[[i]]) + prior[[i]])) %>%
  mutate(machine = list(c("knecht1", "knecht2"), c("knecht3"))) %>%
  mutate(fits = lapply(seq_along(x), function(i) runbg(mstrust(obj[[i]], pars[[i]]), machine = machine[[i]]) ))
# wait a little
dMod.frame <- dMod.frame %>%
  mutate(fits = lapply(seq_along(x), function(i) fits[[i]][["get"]]()[[machine[[i]]]] )) %>%
  mutate(fits = lapply(fits, as.parframe)) %>%
  mutate(plots = lapply(seq_along(x), function(i) plotCombined(prd(mytimes, as.parvec(fits[[i]]), data = data[[i]])) ))
}
\keyword{datasets}
