% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_wrappers.R
\name{dMod_saveMstrust}
\alias{dMod_saveMstrust}
\title{Title}
\usage{
dMod_saveMstrust(fit, path, identifier = "1", FLAGoverwrite = FALSE)
}
\arguments{
\item{fit}{}

\item{path}{}

\item{identifier}{}

\item{FLAGoverwrite}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
