% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{do_checkSensitivities}
\alias{do_checkSensitivities}
\title{Do Check sensitivities over a list of prds}
\usage{
do_checkSensitivities(prds, times, pars, path = "checkSensitivities.pdf")
}
\arguments{
\item{prds}{list of predicition funcitons}

\item{times}{vector of times}

\item{pars}{pars}

\item{path}{filepath for the pdf}
}
\description{
Do Check sensitivities over a list of prds
}
\seealso{
[checkSensitivities()]
}
\concept{doFunctions}
