% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{docpath2dirpath}
\alias{docpath2dirpath}
\title{remove the document name from a document path}
\usage{
docpath2dirpath(doc)
}
\arguments{
\item{doc}{can either be  rstudioapi::getActiveDocumentContext() or a path}
}
\value{
the path with the document removed
}
\description{
remove the document name from a document path
}
