% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{duplicated_values}
\alias{duplicated_values}
\title{Values which are duplicated}
\usage{
duplicated_values(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector of duplicated values
}
\description{
Values which are duplicated
}
\examples{
x <- c(rep(1:10), c(1,3,4,2,1,1,2))
duplicated_values(x)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
