% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{eqnlist_addDefaultCompartment}
\alias{eqnlist_addDefaultCompartment}
\title{Title}
\usage{
eqnlist_addDefaultCompartment(equationList, compName)
}
\arguments{
\item{compName}{}

\item{el}{}
}
\value{

}
\description{
Title
}
