% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{evaluate_formals}
\alias{evaluate_formals}
\title{Load the default values of formals of a function into the Global environment}
\usage{
evaluate_formals(..., indices = 1:length(formals(...)))
}
\arguments{
\item{...}{the function}

\item{indices}{the indices of the formals you want to evaluate}
}
\description{
Load the default values of formals of a function into the Global environment
}
\details{
Useful for debugging
}
\examples{
evaluate_formals(lm, indices = 1:5)
}
