% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{extract_derivs}
\alias{extract_derivs}
\title{Augmentation of getDerivs to subset wrt to which_states.which_pars}
\usage{
extract_derivs(prediction, which_states = NULL, which_pars = NULL)
}
\arguments{
\item{prediction}{a prdframe which contains derivatives as an attribute}

\item{which_states}{which states shall be taken the derivative of ...}

\item{which_pars}{... and with regard to which pars?}
}
\description{
Augmentation of getDerivs to subset wrt to which_states.which_pars
}
