% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{extract_for}
\alias{extract_for}
\alias{extract_apply}
\alias{extract_loopargs}
\title{Get the arguments of a for loop or of lapply}
\usage{
extract_for(textline)

extract_apply(textline)

extract_loopargs(textline)
}
\arguments{
\item{textline}{Line of code}
}
\value{
list(loopvar = "loopingvariable", loopval = "list:ofValues")
}
\description{
Get the arguments of a for loop or of lapply
}
\examples{
extract_loopargs("lapply(names(alpha), function(x) 1)")
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
