% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{extract_variable}
\alias{extract_variable}
\title{Assign code into a variable}
\usage{
extract_variable()
}
\value{
NULL. Modifies the document
}
\description{
Assign code into a variable
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
