% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{fit2obj}
\alias{fit2obj}
\title{Turn a fit into a 2nd order taylor approximation}
\usage{
fit2obj(fit)
}
\arguments{
\item{fit}{output from trust: list with value gradient hessian and argument}
}
\value{
objfun
}
\description{
Turn a fit into a 2nd order taylor approximation
}
