% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{fitErrorModel_plot}
\alias{fitErrorModel_plot}
\title{Get the nice plot of fitErrormodel}
\usage{
fitErrorModel_plot(data)
}
\arguments{
\item{data}{data which goes into fitErrormodel}
}
\description{
Get the nice plot of fitErrormodel
}
\details{
This function is mainly to print the plotting function and to copy/paste it for manipulating it in a script
}
