% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{flexZirkel}
\alias{flexZirkel}
\title{Flexzirkel Stoppuhr}
\usage{
flexZirkel(t1 = 15, t2 = 10, t3 = 10, t4 = 7, t5 = 5)
}
\description{
"coin" = trainieren, "work complete" = aufhö?ren, "bing" = gleich gehts wieder los
}
