% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_compartments}
\alias{get_compartments}
\title{Read COMPARTMENTS}
\usage{
get_compartments(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
data.frame with columns c("name", "unit_type", "unit", "unit_plot_name", "eqn")
}
\description{
Maybe the return value should be just an eqnvec vector c(name = eqn).
}
