% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_description}
\alias{get_description}
\title{Read DESCTRIPTION}
\usage{
get_description(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\description{
Read DESCTRIPTION
}
