% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_doseresponse}
\alias{get_doseresponse}
\title{Read DOSERESPONSE}
\usage{
get_doseresponse(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
Character vector with regex names to be used by str_replace().
}
\description{
Basically copied from get_predictor. Don't know if it's actually needed or what it's cool for.
}
