% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_errors}
\alias{get_errors}
\title{Read ERRORS}
\usage{
get_errors(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
eqnvec.
}
\description{
The returned "errors" eqnvec can be used to make an observation function for the error-model
The "f"-argument to the call of Y can then be c(f, obs).
}
