% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_fitted_pars_from_readme}
\alias{get_fitted_pars_from_readme}
\title{Get fitted pars from Readme.txt}
\usage{
get_fitted_pars_from_readme(readme_file)
}
\arguments{
\item{readme_file}{}
}
\description{
If a Readme.txt file is available, it usually provides fitted parameters. Extract them.
}
\details{
This function is written rather dirtily
}
