% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_inputs}
\alias{get_inputs}
\title{Read INPUTS}
\usage{
get_inputs(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
NULL. It is not yet ready
}
\description{
Read INPUTS
}
