% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_observables}
\alias{get_observables}
\title{Read OBSERVABLES}
\usage{
get_observables(def_content, derived = get_derived(def_content))
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}

\item{derived}{get_derived(def_content) - data.frame with mandatory columns c("name", "eqn")}
}
\value{
eqnvec with the observables
}
\description{
The data in the def_content file contains some columns which I'm not sure are needed in dMod
Therefore, this function just returns an eqnvec with the observables.
}
