% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_doPlots.R
\name{get_parcolindices}
\alias{get_parcolindices}
\title{Get indices of parmaeter columns in a parframe}
\usage{
get_parcolindices(myparframe)
}
\arguments{
\item{myparframe}{.}
}
\description{
Get indices of parmaeter columns in a parframe
}
