% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2d2dMod.R
\name{get_predictor}
\alias{get_predictor}
\title{Read PREDICTOR}
\usage{
get_predictor(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
Character vector with regex names to be used by str_replace().
}
\description{
The predictor in d2d is often called "t", whereas in dMod it is called "time".
Therefore, in input functions, where "t" is explicitly called, the d2d-predictor has to be replaced by the dMod-predictor.
}
