% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2d2dMod.R
\name{get_random}
\alias{get_random}
\title{Read RANDOM}
\usage{
get_random(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
Character vector with parameter names which are condition specific
}
\description{
The conditions section returns one condition in the dMod-sense. It basically returns a trafo with the parameters fixed/changed by this condition
}
