% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2d2dMod.R
\name{get_reactions_or_odes}
\alias{get_reactions_or_odes}
\title{Read REACTIONS}
\usage{
get_reactions_or_odes(
  def_content,
  derived = get_derived(def_content),
  compartments = get_compartments(def_content),
  states = get_states(def_content),
  inputs = get_inputs(def_content)
)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}

\item{derived}{get_derived(def_content) - data.frame with mandatory columns c("name", "eqn")}

\item{compartments}{get_compartments(def_content) - data.frame with mandatory columns c("name", "eqn")}

\item{states}{get_states(def_content) - data.frame with mandatory columns c("name", "compartments")}
}
\value{
eqnlist of all reactions
}
\description{
Read REACTIONS
}
