% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2d2dMod.R
\name{get_states}
\alias{get_states}
\title{Read STATES}
\usage{
get_states(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
data.frame with names c("name", "unit_type", "unit", "unit_plot_name", "compartment", "plot_flag", "plot_name", "is_positive")
}
\description{
Maybe the return value could be just be a named vector c(name = compartment).
}
