% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolsDenial.R
\name{get_substitutions}
\alias{get_substitutions}
\title{Read SUBSTITUTIONS}
\usage{
get_substitutions(def_content)
}
\arguments{
\item{def_content}{The def_content data.frame returned by read_def_content}
}
\value{
eqnvec containing the substitutions.
}
\description{
The substitutions are inserted iteratively. 
Since I don't know in which order, 
I botched a little by running resolveRecurrence(rev(substitutions)) a few times.
}
