% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getFacets}
\alias{getFacets}
\title{Title}
\usage{
getFacets(fp, type)
}
\arguments{
\item{fp, type}{see code of \code{\link[=getPaginateInfo]{getPaginateInfo()}}}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
