% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{getParameters.parframe}
\alias{getParameters.parframe}
\title{get Parameter names of a parframe}
\usage{
\method{getParameters}{parframe}(x)
}
\arguments{
\item{x}{parframe}
}
\description{
get Parameter names of a parframe
}
