% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getParInfo}
\alias{getParInfo}
\title{Extract parInfo data.table from equationlist}
\usage{
getParInfo(equationList)
}
\arguments{
\item{el}{equationList}
}
\value{
data.table
}
\description{
Extract parInfo data.table from equationlist
}
