% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{getSteps}
\alias{getSteps}
\title{getSteps}
\usage{
getSteps(myparframe, nsteps = 5, tol = 1)
}
\arguments{
\item{myparframe}{parframe, result from mstrust}

\item{tol}{tolerance for stepdetection}

\item{min_stepzise}{minimum number of fits landed on this value}
}
\value{
parframe with only these steps
}
\description{
getSteps
}
