% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getUnitInfo}
\alias{getUnitInfo}
\title{Title}
\usage{
getUnitInfo(unitName = NULL, unitKind = NULL, exponent = NULL)
}
\arguments{
\item{unitName}{}

\item{unitKind}{list}

\item{exponent}{list}

\item{unitNameSubset}{}
}
\value{

}
\description{
Title
}
