% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversetools.R
\name{ggshave}
\alias{ggshave}
\title{ggsave with automatically saving metainformation of how the plot was produced}
\usage{
ggshave(filename, plot = last_plot(), device = NULL, path = NULL,
  scale = 1, width = NA, height = NA, units = c("in", "cm", "mm"),
  dpi = 300, limitsize = TRUE, ..., check_overwrite = T)
}
\arguments{
\item{check_overwrite}{Asks before overwriting a plot}

\item{all}{taken from ggsave()}
}
\description{
ggsave with automatically saving metainformation of how the plot was produced
}
