% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gitam}
\alias{gitam}
\title{gitcom --amend from R command line}
\usage{
gitam(string)
}
\arguments{
\item{string}{commit message}
}
\description{
gitcom --amend from R command line
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
