% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gitcom}
\alias{gitcom}
\title{gitcom from R command line}
\usage{
gitcom(string)
}
\arguments{
\item{string}{commit message}
}
\value{

}
\description{
gitcom from R command line
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
