% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{guess_function}
\alias{guess_function}
\title{Guess function name of interest}
\usage{
guess_function(textline)
}
\arguments{
\item{textline}{character(1L), line of script with buggy function}
}
\value{
extracted function name
}
\description{
Guess function name of interest
}
\examples{
textline <- "rng <- rstudioapi::document_range("
textline <- "rstudioapi::document_range("
textline <- "document_range(kjshdf = bla)"
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
