% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{hierarchical_trust}
\alias{hierarchical_trust}
\title{Run hierarchical optimization}
\usage{
hierarchical_trust(dMod.frame, hypothesis = 1, analytic_parms)
}
\arguments{
\item{dMod.frame}{a dMod frame}

\item{hypothesis}{the hypotheses}
}
\description{
Inspired by Daniel Weindl's poster on SBMC 2018
}
\examples{
\dontrun{
source("~/Promotion/Software/dMod/inst/examples/example_dMod.frame/setup.R")
myframe1 <- myframe1 \%>\% appendObj
}
}
