% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{id_test}
\alias{id_test}
\title{Clemens Kreutz Quick identifiability test}
\usage{
id_test(model, hypothesis = 1, r_test = log(10), thresh = 1)
}
\arguments{
\item{model}{dMod.frame}

\item{hypothesis}{which hypothesis}

\item{r_test}{Radius which should be tested}

\item{thresh}{threshold}
}
\value{
table with objective function values, the fit which is performed during this test and the parameter values.
Additional column "identifiable" which is true if the objective function value has been pushed to a value higher than the threshold
}
\description{
Clemens Kreutz Quick identifiability test
}
\examples{
 \dontrun{DL: To do.}
}
