% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{insert2}
\alias{insert2}
\title{Insert only for a subset of conditions}
\usage{
insert2(trafo, expr, condition_sub = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Insert only for a subset of conditions
}
