% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{insert_debugonce}
\alias{insert_debugonce}
\title{Insert debugonce(function)}
\usage{
insert_debugonce()
}
\value{
NULL. Modifies the document
}
\description{
rng <- rstudioapi::document_range(...)
}
\details{
gets turned into

debugonce(rstudioapi::document_range)
rng <- rstudioapi::document_range(...)
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
