% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_fit_analysis}
\alias{insert_fit_analysis}
\title{Insert some common plots after mstrust into your Rmarkdown file}
\usage{
insert_fit_analysis(dMod.frame = "model")
}
\arguments{
\item{dMod.frame}{The dMod.frame you want to analyze}
}
\description{
Insert some common plots after mstrust into your Rmarkdown file
}
