% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_in_document.R
\name{insert_function}
\alias{insert_function}
\title{inserts function assignment of an r-function into the sript}
\usage{
insert_function(f)
}
\arguments{
\item{f}{a function as name}
}
\value{
called for its side effect
}
\description{
inserts function assignment of an r-function into the sript
}
