% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_runbg}
\alias{insert_runbg}
\title{Title}
\usage{
insert_runbg(job_name = "runbg_job", job_type = NULL, dMod.frame = NULL,
  tol = 1, nsteps = 3)
}
\arguments{
\item{job_name}{character with "_job" at the end}

\item{job_type}{character, options: fit, profile, profile_steps}

\item{dMod.frame}{character "dMod.frame"}
}
\description{
Title
}
