% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_seq_along_x_function_i}
\alias{insert_seq_along_x_function_i}
\title{Addin to insert map(seq_along(x), function(i))}
\usage{
insert_seq_along_x_function_i()
}
\description{
Addin to insert map(seq_along(x), function(i))
}
