% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_tee_operator}
\alias{insert_tee_operator}
\title{Addin to insert tee-operator}
\usage{
insert_tee_operator()
}
\description{
Addin to insert tee-operator
}
\seealso{
Other insertfunctions: \code{\link{insert_tee_print}},
  \code{\link{print_r.eqnlist2addReaction_pipe}},
  \code{\link{split_chunks}}
}
