% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_tee_print}
\alias{insert_tee_print}
\title{Addin to insert tee-operator}
\usage{
insert_tee_print()
}
\description{
Addin to insert tee-operator
}
\seealso{
Other insertfunctions: \code{\link{insert_tee_operator}},
  \code{\link{insert_unlink}},
  \code{\link{print_r.eqnlist2addReaction_pipe}},
  \code{\link{split_chunks}}
}
\concept{insertfunctions}
