% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_unit_test}
\alias{insert_unit_test}
\title{Insert Wolfgang's unit test template}
\usage{
insert_unit_test(context = "Context", description = "test_what")
}
\arguments{
\item{context}{The context}

\item{description}{The description in test_that(description)}
}
\description{
Insert Wolfgang's unit test template
}
