% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{insert_values}
\alias{insert_values}
\title{Insert values from another vector with some shared names}
\usage{
insert_values(vec, values)
}
\arguments{
\item{values}{}
}
\value{
the modified vector
}
\description{
Insert values from another vector with some shared names
}
\examples{
vec <- letters[1:3] \%>\% are_names_of(0)
vals <- letters[2:4] \%>\% are_names_of(1)
insert_values(vec, vals)
}
