% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-full_setup.R
\name{install_cfreservedWords}
\alias{install_cfreservedWords}
\title{Copy a standard setup_options_IQRtools.R for custom settings}
\usage{
install_cfreservedWords()
}
\description{
Copy a standard setup_options_IQRtools.R for custom settings
}
