% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-full_setup.R
\name{install_cfrstudio}
\alias{install_cfrstudio}
\title{Quick setup of RStudio on a new machine}
\usage{
install_cfrstudio(theme_name = "pastel on dark")
}
\arguments{
\item{theme_name}{default: "pastel on dark"}
}
\value{
called for side-effect
}
\description{
Quick setup of RStudio on a new machine
}
