% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-snippets.R
\name{install_cfsnippets}
\alias{install_cfsnippets}
\title{Install snippets}
\usage{
install_cfsnippets()
}
\value{
called for its side-effects. Restart RStudio afterwards
}
\description{
Install snippets
}
