% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{install_IQRsnippets}
\alias{install_IQRsnippets}
\title{Install snippets}
\usage{
install_IQRsnippets()
}
\value{
called for its side-effects. Restart RStudio afterwards
}
\description{
Install snippets
}
