% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-full_setup.R
\name{install_thunarshortcuts}
\alias{install_thunarshortcuts}
\title{Install thunar shortcuts}
\usage{
install_thunarshortcuts(FLAGoverwrite)
}
\arguments{
\item{FLAGoverwrite}{}
}
\value{

}
\description{
Install thunar shortcuts
}
