% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jODE.R
\name{jODE_indify_est_grid}
\alias{jODE_indify_est_grid}
\title{Create an index-based version of an est_grid}
\usage{
jODE_indify_est_grid(est_grid, est_vec)
}
\arguments{
\item{est_grid, est_vec}{as in IQRtools, but est_vec should be ordered}
}
\value{
an est_grid with indices instead of names
}
\description{
Used mainly internally in jODE_prepare_matrices
}
