% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jODE.R
\name{jODE_normL2}
\alias{jODE_normL2}
\title{normL2 for julia interface}
\usage{
jODE_normL2(
  est_grid,
  fixed_grid,
  est_vec,
  trafo,
  odes,
  obs,
  err,
  jODE_sourcefile = tempfile("jODE_source_file", fileext = ".jl"),
  data,
  sigma = ":sigma_1"
)
}
\arguments{
\item{est_grid, fixed_grid, est_vec}{as in IQRtools}

\item{trafo, odes, obs, err}{symbolic definitions of functions as in dMod}

\item{jODE_sourcefile}{the intermediate file in which the julia-code is written}

\item{data}{filepath or data.frame}

\item{sigma}{character to paste into julia command. must be one of ":sigma_1" (from data or errormodel) or ":sigma" (from errormodel or data) [] Check which is which}
}
\value{
objective function
}
\description{
normL2 for julia interface
}
