% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jODE.R
\name{jODE_prepare_matrices}
\alias{jODE_prepare_matrices}
\title{Turn fixed_grid and est_grid in numerical matrices}
\usage{
jODE_prepare_matrices(est_grid, fixed_grid, est_vec)
}
\arguments{
\item{est_grid, fixed_grid, fixed_grid}{as in IQRtools}
}
\value{
list of est_mat and fixed_mat

est_mat will be indified \link{jODE_indify_est_grid}
}
\description{
Turn fixed_grid and est_grid in numerical matrices
}
