% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jODE.R
\name{jODE_setup}
\alias{jODE_setup}
\title{Setup Julia for the session and load some jODE functions which don't change}
\usage{
jODE_setup(juliaPath = "/home/daniel/Julia/julia-1.1.1/bin")
}
\arguments{
\item{juliaPath}{path to julia installation}
}
\description{
Setup Julia for the session and load some jODE functions which don't change
}
