% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jODE.R
\name{jODE_write_sourcefile}
\alias{jODE_write_sourcefile}
\title{Conveniencefunction to output source file without having to know the data yet}
\usage{
jODE_write_sourcefile(
  est_grid,
  fixed_grid,
  est_vec,
  trafo,
  odes,
  obs,
  err,
  jODE_sourcefile = tempfile("jODE_source_file", fileext = ".jl")
)
}
\arguments{
\item{est_grid, fixed_grid, fixed_grid}{as in IQRtools}

\item{trafo, odes, obs, err}{symbolic functions as in dMod}

\item{jODE_sourcefile}{path where julia file shall be written}
}
\value{
called for side effect. source file path is returned invisibles
}
\description{
Conveniencefunction to output source file without having to know the data yet
}
