% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jODE.R
\name{jODE_writeTestScript}
\alias{jODE_writeTestScript}
\title{Output a little test script to play around with current objects}
\usage{
jODE_writeTestScript(est_mat, fixed_mat, trafo, odes, obs, err, est_vec)
}
\arguments{
\item{est_mat, fixed_mat}{output of \code{\link{jODE_prepare_matrices}}}

\item{trafo, odes, obs, err}{symbolic definitions}

\item{est_vec}{sorted est_vec}
}
\value{
string which you can use to write a file
}
\description{
Output a little test script to play around with current objects
}
