% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{length_na}
\alias{length_na}
\title{Count missing values in a vector}
\usage{
length_na(vector, missing = TRUE)
}
\arguments{
\item{vector}{The vector which values are to be counted.}

\item{missing}{Boolean. Whether to count missing or non-missing values.}
}
\value{
An integer with the count of values
}
\description{
Count missing values in a vector
}
\examples{

iris$Sepal.Length[4] <- NA
length_na(iris$Sepal.Length)
length_na(iris$Sepal.Length, missing = FALSE)

}
