% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{length_which}
\alias{length_which}
\title{Count values that satisfy a condition}
\usage{
length_which(condition)
}
\arguments{
\item{condition}{The condition for which values are to be counted. Must be
type logical}
}
\value{
An integer with the count of values that satisfy the condition.
}
\description{
Count values that satisfy a condition
}
