% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{mat_simplify}
\alias{mat_simplify}
\alias{mat_denom}
\alias{mat_numer}
\alias{mat_replace}
\title{Simplify with Mathematica}
\usage{
mat_simplify(string)

mat_denom(string)

mat_numer(string)

mat_replace(string, replacements)
}
\arguments{
\item{string}{named vector with equations, not too complicated please. # [] For Mathematica, replacements of sin(x) -> Sin[x] still need to be done}

\item{FLAGsimplify}{call yacas::Simplify or rely solely on the conversion string -> yacas -> string for the simplification}

\item{FLAGverbose}{print input}
}
\value{
named vector with simplified equations
}
\description{
Escapes _ and puts everything into yacas/Mathematica for simplification
}
\details{
yac_simplify (currently disabled) needs yacas and Ryacas installed. mat_simplify needs wolframscript and Matrhematica installed
}
