% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversetools.R
\name{matrix_2_wide_df}
\alias{matrix_2_wide_df}
\title{turn a matrix into a wide data.frame}
\usage{
matrix_2_wide_df(mat, nm)
}
\arguments{
\item{mat}{a matrix}

\item{nm}{vector of names for the matrix elements}
}
\description{
turn a matrix into a wide data.frame
}
\examples{
matrix(1:4, nrow = 2) \%>\% matrix_2_wide_df(letters[1:4])
}
