% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{meld_functions}
\alias{meld_functions}
\title{meld-diff two functions}
\usage{
meld_functions(function1, function2, filenames = "dummymeldcomparison",
  do_unlink = T)
}
\arguments{
\item{function1, function2}{}
}
\value{
side-effect
}
\description{
meld-diff two functions
}
\examples{
meld_functions(lm, glm)
}
\seealso{
Other meld-functions: \code{\link{meld_strings}}
}
\concept{meld-functions}
