% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{meld_strings}
\alias{meld_strings}
\title{Compare two strings in meld}
\usage{
meld_strings(string1, string2, string3 = NULL,
  filenames = "dummymeldcomparison", do_unlink = T)
}
\arguments{
\item{string1, string2, string3}{strings to compare}

\item{filenames}{dummy filenames}

\item{do_unlink}{remove the files}
}
\value{
side-effect
}
\description{
Compare two strings in meld
}
\examples{
meld_functions(lm, glm)
}
\seealso{
Other meld-functions: \code{\link{meld_functions}}
}
\concept{meld-functions}
