% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversetools.R
\name{merge_col2_into_col1}
\alias{merge_col2_into_col1}
\title{Title}
\usage{
merge_col2_into_col1(.data, col1, col2)
}
\arguments{
\item{col2}{}
}
\description{
Title
}
\examples{
df <- data.frame(a = c(1:2, rep(NA, 2), 1:2), b = c(rep(NA,2), 3:4, 3:4))
df \%>\% merge_col2_into_col1(a,b)
df \%>\% merge_col2_into_col1(b,a)
}
