% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{minsert}
\alias{minsert}
\title{Insert parameter values from a given table of covariates}
\usage{
minsert(trafo, pars_to_insert)
}
\arguments{
\item{pars_to_insert}{}
}
\description{
Why "m"insert? - multiple-insert
}
