% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{msNarrow}
\alias{msNarrow}
\title{Construct a parframe with tightly sampled parameters around other parameters stored in a parframe}
\usage{
msNarrow(myparframe, n = 5, sd = 0.2, seed = NULL)
}
\arguments{
\item{myparframe}{a parfame}

\item{n}{how many new rows should be created per row?}

\item{sd, seed}{going to rnorm in msParframe}
}
\value{
a parframe without metanames
}
\description{
Construct a parframe with tightly sampled parameters around other parameters stored in a parframe
}
\examples{
msNarrow(msParframe(c(a = 0, b = 2), n = 2, sd = 3), sd = 0.1)
}
