% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypublish.R
\name{mypublish.eqnvec}
\alias{mypublish.eqnvec}
\title{Print the odes of an eqnvec in a tex and a docx table}
\usage{
\method{mypublish}{eqnvec}(x)
}
\arguments{
\item{x}{eqnlist}
}
\description{
Print the odes of an eqnvec in a tex and a docx table
}
