% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypublish.R
\name{mypublish}
\alias{mypublish}
\title{Standardized outputs for different dMod-Objects}
\usage{
mypublish(x, ...)
}
\arguments{
\item{x}{Model which should be published}

\item{...}{Arguments going to methods}
}
\description{
Standardized outputs for different dMod-Objects
}
