% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypublish.R
\name{mypublish_matlist2wordmath}
\alias{mypublish_matlist2wordmath}
\title{Print a list of matrices as matrices in word math environemt}
\usage{
mypublish_matlist2wordmath(x, filename)
}
\arguments{
\item{x}{a list of matrices}

\item{filename}{filename without ending}
}
\description{
Print a list of matrices as matrices in word math environemt
}
