% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypublish.R
\name{mypublish_matlist2wordtable}
\alias{mypublish_matlist2wordtable}
\title{Print a list of matrices to a single word file with a table of these matrices}
\usage{
mypublish_matlist2wordtable(x, path, digits = 2)
}
\arguments{
\item{x}{list of matrices}

\item{path}{filename without ending}

\item{digits}{digits for rounding}
}
\description{
Print a list of matrices to a single word file with a table of these matrices
}
\examples{
\dontrun{
list(r_0 = matrix(1:9, ncol = 3),
     r_1 = matrix(0,ncol = 3, nrow = 3),
     r_opt_1 = matrix(2:10, nrow = 3)) %>%
  imap( ~.x %>% round(2) %>%
          xtable(align = rep("",4)) %>%
          print(tabular.environment="pmatrix",
                include.colnames = F,
                include.rownames = F,
                floating = F,
                hline.after = NULL)) %>%
          mypublish_mat("matrices")
}
}
