% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_values.R
\docType{data}
\name{myvec}
\alias{myvec}
\alias{mydf}
\alias{mymat}
\alias{mytibble}
\title{Useful values}
\format{An object of class \code{integer} of length 3.}
\usage{
myvec

mydf

mymat

mytibble
}
\description{
Useful values
}
\keyword{datasets}
