% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{obj_condition_wise}
\alias{obj_condition_wise}
\title{Evaluate an objective function condition-wise}
\usage{
obj_condition_wise(obj, pars, constr1 = NULL, constr2 = NULL,
  conditions = NULL, ...)
}
\arguments{
\item{obj}{objfn}

\item{conditions}{vector of conditions to be evaluated. if null, all conditions are evaluated}

\item{pouter}{pars at which obj is evaluated}
}
\value{
A tibble with rows corresponding to conditions and cols to several results of the objective function, such as value, gradient...
}
\description{
Easily see which condition contributes how much.
Be careful to use this in an actual calculation, as in adding the results back together, because then the prior is evaluated multiple times
}
