% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{objlist_template}
\alias{objlist_template}
\title{Create an empty objlist full of zeros}
\usage{
objlist_template(pars)
}
\arguments{
\item{pars}{a named vector, names will be used to construct the gradient and hessian}
}
\value{
objlist full of zeros
}
\description{
Create an empty objlist full of zeros
}
\examples{
objlist_template(structure(1:3, names = letters[1:3]))
}
