% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{open_all_scripts_in_dir}
\alias{open_all_scripts_in_dir}
\title{Open all scripts of a folder in rstudio}
\usage{
open_all_scripts_in_dir(dirname = "Scripts", pattern = "\\\\.R$")
}
\description{
Open all scripts of a folder in rstudio
}
