% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{open_in_calc}
\alias{open_in_calc}
\title{Write an excel file and cat a name to open it.}
\usage{
open_in_calc(.x)
}
\arguments{
\item{.x}{data.frame}
}
\description{
Write an excel file and cat a name to open it.
}
