% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{open_in_file}
\alias{open_in_file}
\title{Write a characeter vector to a file and open it.}
\usage{
open_in_file(.x)
}
\arguments{
\item{.x}{vector}
}
\description{
Write a characeter vector to a file and open it.
}
