% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_doPlots.R
\name{optimal_cuts}
\alias{optimal_cuts}
\title{Choose the best out of several cut lenghts}
\usage{
optimal_cuts(vec, possibilities)
}
\arguments{
\item{vec}{the vector you want to cut into pieces}

\item{possibilities}{the possible sizes of the slices in preferred order}
}
\description{
Choose the best out of several cut lenghts
}
\examples{
vec <- letters
possibilities <- c(12,15,10)
optimal_cuts(vec, possibilities)
possibilities <- c(12,10,13)
optimal_cuts(vec, possibilities)
}
