% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{parframes_summary}
\alias{parframes_summary}
\title{Quickly print the number of converged fits}
\usage{
parframes_summary(dMod.frame, hypothesis = NULL, tol = 1)
}
\arguments{
\item{dMod.frame}{dMod.frame with parframes column}

\item{hypothesis}{hypothesis, if null, analyse all hypotheses given in the dMod.frame}

\item{tol}{tol going to \code{dMod:::stepDetect}}
}
\description{
Quickly print the number of converged fits
}
