% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_condition}
\alias{petab_condition}
\title{Constructor for Conditions}
\usage{
petab_condition(conditionId, conditionName = NULL, ...)
}
\arguments{
\item{conditionId}{character Condition ID}

\item{conditionName}{character Condition ID for plotting}

\item{...}{parameterOrSpeciesOrCompartmentId1. Numeric (value) or string (different parameterId)}
}
\value{

}
\description{
https://petab.readthedocs.io/en/stable/documentation_data_format.html#condition-table
}
