% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_files}
\alias{petab_files}
\title{List petab files}
\usage{
petab_files(modelname, path, FLAGTestCase = FALSE)
}
\arguments{
\item{modelname}{}

\item{path}{}

\item{FLAGTestCase}{modelname <- "Boehm_JProteomeRes2014"
path <- "../../Software/dMod/BenchmarkModels/Boehm_JProteomeRes2014/"
FLAGTestCase <- FALSE
path <- "../../Software/dMod/PEtabTests/0001"
FLAGTestCase <- TRUE}
}
\description{
List petab files
}
