% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getMeasurementParsMapping}
\alias{petab_getMeasurementParsMapping}
\title{Parse the parameters columns in measurementData}
\usage{
petab_getMeasurementParsMapping(
  measurementData,
  column = c("observableParameters", "noiseParameters")[1]
)
}
\arguments{
\item{measurementData}{petab$measurementData. Uses only columns observableId,simulationConditionId and "column"}

\item{column}{character(1L). one of c("observableParameters", "noiseParameters")}
}
\value{
data.table(INNERPARAMETER, OUTERPARAMETER)
}
\description{
Parse the parameters columns in measurementData
}
\examples{
measurementData <- data.table(
observableId = c("obsE"),
simulationConditionID = c("C1"),
noiseParameters = "sigmaAbs"
)

petab_getMeasurementParsMapping(measurementData, "noiseParameters")

# Old examples. Need to adapt
# petab_getMeasurementParsMapping(observableId, simulationConditionId, "sigmaAbs;1")
# petab_getMeasurementParsMapping(observableId, simulationConditionId, "sigmaAbs;sigmaRel")
# petab_getMeasurementParsMapping(observableId, simulationConditionId, NULL, "1;offsE")
# petab_getMeasurementParsMapping(rep("obsE", 2), c("C1","C2"), NULL, c("1;offsE", "2;offsE_C2"))
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
