% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_joinDataCondObs}
\alias{petab_joinDataCondObs}
\title{Create one big table containing measurementData, observables and experimentalCondition}
\usage{
petab_joinDataCondObs(petab)
}
\arguments{
\item{petab}{\link{petab}}
}
\value{

}
\description{
Create one big table containing measurementData, observables and experimentalCondition
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
