% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_lint}
\alias{petab_lint}
\title{Title}
\usage{
petab_lint(petab)
}
\arguments{
\item{petab}{}
}
\value{
list of errors
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
