% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_measurementData}
\alias{petab_measurementData}
\title{Constructor for Measurements}
\usage{
petab_measurementData(
  observableId,
  simulationConditionId,
  measurement,
  time,
  observableParameters = list(NULL, "1;1", "scale_i;offset_i")[[1]],
  noiseParameters = list(1, "error_obsi_ADD;error_obsi_REL")[[1]],
  datasetId = NULL,
  replicateId = NULL,
  preequilibrationConditionId = NULL
)
}
\arguments{
\item{observableId}{}

\item{simulationConditionId}{}

\item{measurement}{}

\item{time}{}

\item{observableParameters}{numeric string or null}

\item{noiseParameters}{numeric, string or null: Measurement noise or parameter name}

\item{datasetId}{}

\item{replicateId}{}

\item{preequilibrationConditionId}{}
}
\description{
Constructor for Measurements
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
