% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_modelname_path}
\alias{petab_modelname_path}
\title{Title}
\usage{
petab_modelname_path(filename)
}
\arguments{
\item{filename}{path ending in .petab}
}
\value{
list(modelname, path)
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
