% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_mutateDCO}
\alias{petab_mutateDCO}
\title{Consistently mutate the dco}
\usage{
petab_mutateDCO(pe, i, j)
}
\arguments{
\item{pe}{[petab()]}

\item{i}{expression to subset the dco}

\item{j}{Must be a call of the form `:=`(hsdhfgjs)}
}
\value{

}
\description{
If j is supplied, the full dco is always returned!
So, subsetting might need two steps
}
