% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_observables}
\alias{petab_observables}
\title{Constructor for Observables}
\usage{
petab_observables(
  observableId,
  observableName = NULL,
  observableFormula = "observableParameter${n}_${observableId} * state1",
  observableTransformation = c("lin", "log", "log10")[[1]],
  noiseFormula = c(1, "noiseParameter${n}_${observableId}")[[1]],
  noiseDistribution = c("normal", "laplace")[[1]]
)
}
\arguments{
\item{observableId}{}

\item{observableName}{}

\item{observableFormula}{}

\item{observableTransformation}{}

\item{noiseFormula}{}

\item{noiseDistribution}{}
}
\value{

}
\description{
Constructor for Observables
}
