% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_parameters}
\alias{petab_parameters}
\title{Constructor for Parameters}
\usage{
petab_parameters(
  parameterId,
  parameterName = NULL,
  parameterScale = c("log", "lin", "log10")[[1]],
  lowerBound = 0,
  upperBound = 10000,
  nominalValue = 1,
  estimate = c(1, 0)[[1]],
  initializationPriorType = c("parameterScaleUniform", "uniform", "normal", "laplace",
    "logNormal", "logLaplace", "parameterScaleNormal", "parameterScaleLaplace")[[1]],
  initializationPriorParameters = "-1;1",
  objectivePriorType = c("parameterScaleNormal", "parameterScaleUniform", "uniform",
    "normal", "laplace", "logNormal", "logLaplace", "parameterScaleLaplace")[[1]],
  objectivePriorParameters = "-1;1"
)
}
\arguments{
\item{parameterId}{}

\item{parameterName}{}

\item{parameterScale}{}

\item{lowerBound}{}

\item{upperBound}{}

\item{nominalValue}{}

\item{estimate}{}

\item{initializationPriorType}{}

\item{initializationPriorParameters}{}

\item{objectivePriorType}{}

\item{objectivePriorParameters}{}
}
\value{

}
\description{
Constructor for Parameters
}
