% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_plotData}
\alias{petab_plotData}
\title{Plot Petab data}
\usage{
petab_plotData(
  petab,
  conditionId = NULL,
  observableId = NULL,
  datasetId = NULL,
  aeslist = NULL,
  FLAGUseObservableTransformation = TRUE,
  FLAGmeanLine = TRUE,
  ggCallback = facet_wrap_paginate(~observableId, nrow = 4, ncol = 4, scales = "free")
    + geom_blank(),
  filename = NULL,
  FLAGfuture = TRUE,
  width = 21,
  height = 29.7,
  scale = 1,
  units = "cm"
)
}
\arguments{
\item{petab}{A \link{petab} object}

\item{conditionId, observableId, datasetId}{Subset data to plot. Find out suitable values with \code{\link[=petab_dput]{petab_dput()}}}

\item{aeslist}{list of formulas used to call \code{\link[ggplot2:aes_]{ggplot2::aes_q()}}.
Default aesthetics are list(x = ~time, y = ~measurement, color = ~conditionId).
Can be overriden by aeslist}

\item{FLAGmeanLine}{draw line connecting the means}

\item{ggCallback}{additional stuff to add to the ggplot, such as a call to \code{\link[ggplots2:labs]{ggplots2::labs()}} or scales}

\item{filename, width, height, scale, units}{see \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}

\item{FLAGfuture}{export asynchronously with the future package}

\item{paginateInfo}{see \code{\link[=cf_paginateInfo]{cf_paginateInfo()}}}
}
\value{
ggplot
}
\description{
Automatic multipage export in the background with the future package and ggforce::facet_wrap_paginate
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
