% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{plotCombined.prdlist}
\alias{plotCombined.prdlist}
\title{Title}
\usage{
\method{plotCombined}{prdlist}(prediction, data = NULL, ...,
  scales = "free", facet = "wrap", transform = NULL,
  aesthetics = NULL)
}
\arguments{
\item{aesthetics}{}
}
\description{
Title
}
