% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{plotCombined}
\alias{plotCombined}
\title{Plot a list of model predictions and a list of data points in a combined plot}
\usage{
plotCombined(prediction, ...)
}
\description{
Plot a list of model predictions and a list of data points in a combined plot
}
