% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{plotCombined.tbl_df}
\alias{plotCombined.tbl_df}
\title{Title}
\usage{
\method{plotCombined}{tbl_df}(model, hypothesis = 1, index = 1, ...,
  plotErrorBands = F)
}
\arguments{
\item{plotErrorBands}{}
}
\description{
Title
}
