% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{plotData.datalist}
\alias{plotData.datalist}
\title{Title}
\usage{
\method{plotData}{datalist}(data, ..., scales = "free", facet = "wrap",
  transform = NULL, aesthetics = NULL)
}
\arguments{
\item{transform}{}
}
\description{
Title
}
