% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{plotMulti.prdfn}
\alias{plotMulti.prdfn}
\title{Title}
\usage{
\method{plotMulti}{prdfn}(x, times, parframe, data = NULL, ...,
  colorAsFactor = F)
}
\arguments{
\item{colorAsFactor}{}
}
\description{
Title
}
