% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{plotPars}
\alias{plotPars}
\alias{plotPars.parframe}
\alias{plotPars.tbl_df}
\title{Plot parameter values for a fitlist}
\usage{
plotPars(x, ...)

\method{plotPars}{parframe}(x, tol = 1, ...)

\method{plotPars}{tbl_df}(dMod.frame, hypothesis = 1, ..., tol = 1)
}
\arguments{
\item{x}{parameter frame as obtained by as.parframe(mstrust)}

\item{...}{arguments for subsetting of x}

\item{tol}{maximal allowed difference between neighboring objective values
to be recognized as one.}

\item{nsteps}{number of steps from the waterfall plot}
}
\description{
Plot parameter values for a fitlist
}
