% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{prepare_argpath}
\alias{prepare_argpath}
\alias{plot_argpath}
\title{Look at the argpath of a fit}
\usage{
prepare_argpath(fit, ncluster = 1)

plot_argpath(argpath)
}
\arguments{
\item{fit}{a fit returned from dMod::trust(..., blather = T)}

\item{argpath}{result from \link{prepare_argpath}}
}
\value{
a data.frame for plotting with plot_argpath
}
\description{
Requires the fit to be run with blather = T.
The same functionality was actually already implemented by Malenka in dMod::plot.parlist(..., path = T)
Both are nice and you see different things.
Maybe it would be cool to implement a clustering-option into Malenkas plot.parlist()
}
