% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{print_mathematica.character}
\alias{print_mathematica.character}
\title{Print a character in mathematica format}
\usage{
print_mathematica.character(mycharacter)
}
\arguments{
\item{mycharacter}{}
}
\description{
Print a character in mathematica format
}
\examples{
 print_mathematica.character(paste0(letters, "_", letters))
}
