% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{print_mathematica.eqnlist}
\alias{print_mathematica.eqnlist}
\title{Print ode vector in mathematica format}
\usage{
print_mathematica.eqnlist(myeqnlist)
}
\arguments{
\item{myeqnlist}{an eqnlist}
}
\description{
Print ode vector in mathematica format
}
