% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{print_mathematica.eqnvec}
\alias{print_mathematica.eqnvec}
\title{Print ode vector in mathematica format}
\usage{
print_mathematica.eqnvec(myeqnvec)
}
\arguments{
\item{myeqnvec}{an eqnvec}
}
\description{
Print ode vector in mathematica format
}
