% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_mathematica.matrix}
\alias{print_mathematica.matrix}
\title{Print mathematica}
\usage{
print_mathematica.matrix(mat, name_in_mathematica = "M")
}
\arguments{
\item{mat}{}
}
\description{
Print mathematica
}
\examples{
print_mathematica.matrix(diag(1:5))
}
