% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{print_mathematica.name_equals_value}
\alias{print_mathematica.name_equals_value}
\title{Print the values such that they are assignments in mathematica}
\usage{
print_mathematica.name_equals_value(vec)
}
\arguments{
\item{vec}{}
}
\description{
Print the values such that they are assignments in mathematica
}
\examples{
print_mathematica.name_equals_value(c(a = 1, b = 2))
}
