% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_r.df2tribble}
\alias{print_r.df2tribble}
\title{Print data.frame as tribble-call}
\usage{
print_r.df2tribble(df, FLAGround = F)
}
\arguments{
\item{FLAGround}{round numeric columns}
}
\value{
insertion into your script
}
\description{
Print data.frame as tribble-call
}
\examples{
print_r.df2tribble(mydf)
}
