% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_r.eqnlist2addReaction_pipe}
\alias{print_r.eqnlist2addReaction_pipe}
\title{Print an eqnlist as a pipe of addReaction()-calls and insert them in place}
\usage{
print_r.eqnlist2addReaction_pipe(reactions)
}
\arguments{
\item{reactions}{}
}
\description{
Print an eqnlist as a pipe of addReaction()-calls and insert them in place
}
\seealso{
Other printtools: \code{\link{print_mathematica.character}},
  \code{\link{print_mathematica.eqnlist}},
  \code{\link{print_mathematica.eqnvec}},
  \code{\link{print_mathematica.matrix}},
  \code{\link{print_mathematica.name_equals_value}},
  \code{\link{print_r.named_vector}}

Other insertfunctions: \code{\link{insert_tee_operator}},
  \code{\link{insert_tee_print}},
  \code{\link{insert_unlink}}, \code{\link{split_chunks}}
}
\concept{insertfunctions}
\concept{printtools}
