% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print_r.named_vector}
\alias{print_r.named_vector}
\title{nicely formatted dput for named vectors}
\usage{
print_r.named_vector(myvec)
}
\arguments{
\item{myvec}{}
}
\description{
nicely formatted dput for named vectors
}
