% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_rscript_comment}
\alias{print_rscript_comment}
\title{Print a string with comments into your script}
\usage{
print_rscript_comment(vec)
}
\arguments{
\item{vec}{}
}
\description{
Print a string with comments into your script
}
