% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_organization.R
\name{proj_create_folders}
\alias{proj_create_folders}
\title{Create the basic project folder structure}
\usage{
proj_create_folders(path = "Work")
}
\arguments{
\item{path}{the path of the project folder}
}
\value{
this function is called for its side-effect
}
\description{
Create the basic project folder structure
}
