% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{protect_so}
\alias{protect_so}
\title{Zip .so files in a file with unique name}
\usage{
protect_so(model)
}
\arguments{
\item{model}{a dMod.frame}
}
\description{
Zip .so files in a file with unique name
}
