% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pullJobs}
\alias{pullJobs}
\title{Git pull all projects in the PROJTOOLS fodler}
\usage{
pullJobs()
}
\description{
Git pull all projects in the PROJTOOLS fodler
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
