% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2d2dMod.R
\name{read_def_content}
\alias{read_def_content}
\title{Read the *.def file an prepare it for later use}
\usage{
read_def_content(connection)
}
\arguments{
\item{connection}{The path to the file}
}
\value{
A data.frame with columns "def_content" and "section". "def_content" are the lines present in "section"
}
\description{
The function determines automatically if it's a data.def or a model.def file by checking if any of the headings REACTIONS, ODES or DERIVED are given. These are unique to model.def files.
}
