% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{readDMod.frame}
\alias{readDMod.frame}
\title{Load a saved dMod.frame}
\usage{
readDMod.frame(filename)
}
\arguments{
\item{filename}{Character, filename e.g. "myframe.rds"

a dMod.frame}
}
\description{
A dMod.frame saved by saveRDS is read and the DLLs required by its fn-objects are loaded as a side-effect
}
