% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{recover_runbg}
\alias{recover_runbg}
\title{Recover a runbg-jog}
\usage{
recover_runbg(machine, filename, wait = FALSE)
}
\arguments{
\item{machine}{What you entered when you started the job}

\item{filename}{What you entered when you started the job (if you entered nothing look at the RData-files that were generated)}

\item{wait}{What you entered when you started the job}
}
\value{
the normal return value of runbg
}
\description{
Recover a runbg-jog when the job is lost somewhere in the abyss between restarting sessions and not saving
}
\author{
Daniel Kaschek, Daniel Lill just selected the respective lines
}
