% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{refactor_functionCall}
\alias{refactor_functionCall}
\title{Title}
\usage{
refactor_functionCall()
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
