% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{relPath_from_1_to_2}
\alias{relPath_from_1_to_2}
\title{Relative path between two absolute paths}
\usage{
relPath_from_1_to_2(from, to)
}
\arguments{
\item{from, to}{absPaths}
}
\value{
relPath
}
\description{
Relative path between two absolute paths
}
\examples{
relPath_from_1_to_2("~/test", "/usr/local")
}
