% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversetools.R
\name{remove_geom}
\alias{remove_geom}
\title{Remove a geom layer}
\usage{
remove_geom(x, geom_type, last_only = T)
}
\arguments{
\item{x}{ggplot object}

\item{geom_type}{Character such as "GeomLine"}

\item{last_only}{only the topmost layer of this kind is removed}
}
\value{
the plot with the layers removed
}
\description{
Remove a geom layer
}
\details{
This function was mainly written by Pedro Aphalo on https://stackoverflow.com/questions/13407236/remove-a-layer-from-a-ggplot2-chart#51612148
I only added the last_only - argument
}
