% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModFrameClass.R
\name{reMutateCalls}
\alias{reMutateCalls}
\title{Redo calls that are saved in the .calls - column}
\usage{
reMutateCalls(dMod.frame, whichCalls)
}
\arguments{
\item{dMod.frame}{A dMod.frame with a record of the calls}

\item{whichCalls}{Numeric indices or character to subset .calls by name}
}
\description{
This guy doesn't work because the calls get wrapped into lists again
}
