% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{rename_script}
\alias{rename_script}
\title{Rename a script, change filename occurences within script, rename output folder}
\usage{
rename_script(from, to)
}
\arguments{
\item{from, to}{as in file.rename}
}
\description{
Rename a script, change filename occurences within script, rename output folder
}
