% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{report_dMod.frame}
\alias{report_dMod.frame}
\title{Print plots and shit}
\usage{
report_dMod.frame(model, folder = tpaste0("report/"))
}
\arguments{
\item{model}{DL to do}

\item{folder}{DL to do}
}
\description{
Print plots and shit
}
