% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{runtime}
\alias{runtime}
\title{The runtime of some code}
\usage{
runtime(...)
}
\arguments{
\item{...}{Some code}
}
\value{
The output of the code with an attribute "runtime"
}
\description{
This function can be used, when you want to benchmark, but also access the results.
}
