% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_addOneCompartment}
\alias{sbml_addOneCompartment}
\title{Title}
\usage{
sbml_addOneCompartment(model, compName, compSize)
}
\arguments{
\item{model}{}

\item{compName}{}

\item{compsize}{}
}
\value{

}
\description{
Title
}
