% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_addOneReaction}
\alias{sbml_addOneReaction}
\title{Title}
\usage{
sbml_addOneReaction(
  model,
  reactionName,
  reversible = 1,
  reactants,
  products,
  equation,
  parName,
  parValue,
  parUnit
)
}
\arguments{
\item{model}{}

\item{reactionName}{}

\item{reversible}{}

\item{reactants}{}

\item{products}{}

\item{equation}{}

\item{parName}{}

\item{parValue}{}

\item{parUnit}{}
}
\value{

}
\description{
Title
}
