% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_addOneSpecies}
\alias{sbml_addOneSpecies}
\title{Title}
\usage{
sbml_addOneSpecies(model, speciesName, compName, initialAmount)
}
\arguments{
\item{model}{}

\item{speciesName}{}

\item{compName}{}

\item{speciesInitialAmount}{}
}
\value{

}
\description{
Title
}
