% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_addOneUnit}
\alias{sbml_addOneUnit}
\title{Add unit}
\usage{
sbml_addOneUnit(
  model,
  unitName = "litre_per_mole_per_second",
  unitKind = c("UNIT_KIND_LITRE", "UNIT_KIND_MOLE", "UNIT_KIND_SECOND"),
  exponent = c(1, -1, -1)
)
}
\arguments{
\item{model}{}

\item{unitName}{}

\item{unitKind}{}

\item{exponent}{}
}
\value{

}
\description{
Add unit
}
