% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_exportEquationList}
\alias{sbml_exportEquationList}
\title{Title}
\usage{
sbml_exportEquationList(
  equationList,
  filename,
  modelname = "Model",
  unitInfo = getUnitInfo(),
  speciesInfo = getSpeciesInfo(equationList),
  parInfo = getParInfo(equationList),
  compartmentInfo = getCompartmentInfo(equationList)
)
}
\arguments{
\item{equationList}{}

\item{filename}{}

\item{modelname}{}

\item{unitInfo}{}

\item{speciesInfo}{}

\item{parInfo}{}

\item{compartmentInfo}{}
}
\value{

}
\description{
Title
}
\examples{

library(conveniencefunctions)
.. Eqnlist and objects -----
el <- NULL
el <- addReaction(el, from = "E + S", to = "ES", rate = "(kon)*E*S",
                  description = "production of complex")
el <- addReaction(el, from = "ES", to = "E + S", rate = "koff*ES",
                  description = "decay of complex")
el <- addReaction(el, from = "ES", to = "E + P", rate = "kcat*ES",
                  description = "production of product")
el <- eqnlist_addDefaultCompartment(el, "cytoplasm")

filename <- file.path(tempdir(), "model.xml")

parInfo <- data.table(tibble::tribble(
  ~parName, ~parValue, ~parUnit,
  "kon"   ,     0.10,"litre_per_mole_per_second" ,
  "koff"  ,     0.55,"per_second" ,
  "kcat"  ,     1.00,"per_second" ))

sbml_exportEquationList(el, filename, parInfo = parInfo)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
