% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_initialize}
\alias{sbml_initialize}
\title{Title}
\usage{
sbml_initialize(modelname = "EnzymaticReaction", sbmlDoc)
}
\arguments{
\item{modelname}{}

\item{sbmlDoc}{}
}
\value{

}
\description{
Title
}
