% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_reactionAddModifiers}
\alias{sbml_reactionAddModifiers}
\title{Title}
\usage{
sbml_reactionAddModifiers(reaction, modifiers)
}
\arguments{
\item{reaction}{}

\item{products}{list(name = c("a", "b"), stoichiometry = c(1,2))}
}
\value{

}
\description{
Title
}
