% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_reactionAddProducts}
\alias{sbml_reactionAddProducts}
\title{Title}
\usage{
sbml_reactionAddProducts(reaction, speciesNames)
}
\arguments{
\item{reaction}{}

\item{speciesNames}{}
}
\value{

}
\description{
Title
}
