% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_reactionAddReactants}
\alias{sbml_reactionAddReactants}
\title{Title}
\usage{
sbml_reactionAddReactants(reaction, reactants)
}
\arguments{
\item{reaction}{}

\item{speciesNames}{}
}
\value{

}
\description{
Title
}
