% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_centerY}
\alias{scale_centerY}
\title{Center y axes of different facets around 0 or 1}
\usage{
scale_centerY(bycols, fun = min, dplot, FLAGcenterLog = TRUE, ...)
}
\arguments{
\item{bycols}{character vector of columns to summarize by. These are typically columns you facet over}

\item{fun}{min or max}

\item{dplot}{data.table(measurement = nominal values, bycols...). Typically, the data that went into the plot}

\item{FLAGcenterLog}{do the centering on log scale}

\item{...}{arguments to geom_blank}
}
\value{
geom_blank
}
\description{
solution adapted from https://stackoverflow.com/questions/31782799/how-to-use-free-scales-but-keep-a-fixed-reference-point-in-ggplot
}
\details{
You need to call this function twice, once with min and once with max
}
\examples{

# linear scale
dx <- copy(mtcars)
dx <- data.table(dx)
dx[,`:=`(measurement = wt-mean(wt))]

ggplot(dx, aes(mpg, measurement)) + 
  facet_wrap( ~ cyl, scales = "free_y") + 
  geom_point() +
  scale_centerY("cyl", min, dx, FALSE) +
  scale_centerY("cyl", max, dx, FALSE)

# log scale
dx <- copy(mtcars)
dx <- data.table(dx)
dx[,`:=`(measurement = wt/mean(wt))]

ggplot(dx, aes(mpg, measurement)) + 
  facet_wrap( ~ cyl, scales = "free_y") + 
  geom_point() +
  scale_centerY("cyl", min, dx, TRUE) +
  scale_centerY("cyl", max, dx, TRUE) + 
  scale_y_log10()

}
