% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{send_runbg_mail}
\alias{send_runbg_mail}
\title{Send a mail informing you}
\usage{
send_runbg_mail(job_name, to = "dl140@physik.uni-freiburg.de")
}
\arguments{
\item{job_name}{the name of the runbg_job}
}
\description{
Send a mail informing you
}
