% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_new_chapter}
\alias{set_new_chapter}
\title{Sets new chapter in R script in RStudio}
\usage{
set_new_chapter()
}
\value{
Creates a vertical line in R script
}
\description{
Sets new chapter in R script in RStudio
}
\references{
https://www.statworx.com/de/blog/defining-your-own-shortcut-in-rstudio/
}
