% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{shrink_fits}
\alias{shrink_fits}
\alias{shrink_plots}
\title{Remove cols from dMod.frame}
\usage{
shrink_fits(dMod.frame)

shrink_plots(dMod.frame)
}
\arguments{
\item{dMod.frame}{}
}
\description{
Remove cols from dMod.frame
}
