% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-snippets.R
\name{snippet_remove}
\alias{snippet_remove}
\title{copied from snippr}
\usage{
snippet_remove(
  name,
  language = "r",
  path = "~/.R/snippets/r.snippets",
  error = TRUE
)
}
\description{
copied from snippr
}
