% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{spread_list_column}
\alias{spread_list_column}
\alias{unnest_name}
\title{Unnest a list-column into a key-value pair, if the list entries are named vectors}
\usage{
spread_list_column(myframe, unique = "condition",
  unnest_var = "gradient")

unnest_name(myframe, unique = "condition", unnest_var = "gradient")
}
\arguments{
\item{myframe}{a tibble with list columns}

\item{unique}{Character. A column name with unique elements in each row. \code{myframe} is split up and joined back together,
so a unique identifier is needed to merge the two tibbles back together.}

\item{unnest_var}{Character. The column name of the column which should be spread

 @export}
}
\description{
This function was built to look at the results of obj_condition_wise more easily
}
\examples{
tibble(condition = 1, gradient = list(c(a=1, b=2, c=3))) \%>\% spread_list_column()
}
