% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_align}
\alias{str_align}
\title{Align some lines of code along a pattern}
\usage{
str_align(string, pattern = "<-")
}
\arguments{
\item{string}{string}

\item{pattern}{pattern}
}
\value{
aligned string
}
\description{
Align some lines of code along a pattern
}
\examples{
mystr <- 'a<- 1
sdlfkjslkdjlskdfjlsjdkjslkd<-2'
str_align(mystr)
}
