% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_detect_any}
\alias{str_detect_any}
\title{Is (any element) of pattern in any of the elements of vec?}
\usage{
str_detect_any(vec, pattern)
}
\arguments{
\item{vec}{named vector}

\item{pattern}{regex as in str_subset}
}
\value{
Logical
}
\description{
I usually use it with a pattern of length 1.
}
\examples{
str_detect_any(letters[1:3], c("c"))
str_detect_any(letters[1:3], c("d"))
str_detect_any(letters[1:3], c("c","d"))
}
