% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_escape}
\alias{str_escape}
\title{Escape all special characters in a string}
\usage{
str_escape(string)
}
\arguments{
\item{string}{string}
}
\value{
string with single escapes before special characters (don't be confused by the double escapes from "print")
}
\description{
Escape all special characters in a string
}
\examples{
str_escape("a.(asdfsadf)") \%>\% cat
}
