% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr.R
\name{str_replace_in_file}
\alias{str_replace_in_file}
\title{Title}
\usage{
str_replace_in_file(path, pattern, replacement, walkthrough = F, writeout = F)
}
\arguments{
\item{path}{filepath}

\item{pattern, replacement}{as in str_replace}

\item{walkthrough}{walk through each matching line and decide}

\item{writeout}{write to file}
}
\value{
the modified string
}
\description{
Title
}
