% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_subset_keep_names}
\alias{str_subset_keep_names}
\title{str_subset that preserves names}
\usage{
str_subset_keep_names(vec, pattern)
}
\arguments{
\item{pattern}{}
}
\description{
str_subset that preserves names
}
