% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_subset_name}
\alias{str_subset_name}
\alias{str_subset_name<-}
\title{subset a vector by matching names against a pattern}
\usage{
str_subset_name(vec, pattern)

str_subset_name(vec, pattern) <- value
}
\arguments{
\item{vec}{named vector}

\item{pattern}{regex as in \link{str_subset}}
}
\description{
This function is inspired by str_subset from stringr
}
\examples{
myvec <- c("a", "b", "ab", "ba") \%>\% are_names_of(runif)
myvec \%>\% str_subset_name("^a")
}
