% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr.R
\name{str_subset_not}
\alias{str_subset_not}
\alias{str_subset_keep_names_not}
\title{The "opposite" of str_subset.}
\usage{
str_subset_not(vec, pattern)

str_subset_keep_names_not(vec, pattern)
}
\arguments{
\item{vec}{named vector}

\item{pattern}{regex as in str_subset}
}
\description{
Preserves names
}
