% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{subsetparameters}
\alias{subsetparameters}
\title{Select parameter columns of parframe}
\usage{
subsetparameters(parf, parameters)
}
\arguments{
\item{parf}{parframe}

\item{parameters}{keine ahnung mehr?}
}
\description{
Select parameter columns of parframe
}
