% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{subtract_by_name}
\alias{subtract_by_name}
\title{Subtract elements with matching names}
\usage{
subtract_by_name(.x, .y)
}
\arguments{
\item{.x, .y}{vecs or matrices}
}
\value{
.y-.x at the respective positions
}
\description{
Subtract elements with matching names
}
