% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{tblprint}
\alias{tblprint}
\title{"tibble-print" a matrix}
\usage{
tblprint(x, i = 10, j = 5, drop = T)
}
\arguments{
\item{x}{matrix}

\item{i, j}{how many rows/cols}

\item{drop}{drop}
}
\value{
x without subsetting
}
\description{
print only the first i rows and j cols
}
