% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{test_dMod.frame}
\alias{test_dMod.frame}
\title{Test some basic functionalities of a dMod.frame}
\usage{
test_dMod.frame(model, test_plots = T, test_obj = T)
}
\arguments{
\item{model}{A dMod.frame}

\item{test_plots, test_obj}{logical}
}
\value{
an augmented dMod.frame containing columns named like "tested_*"
}
\description{
Test some basic functionalities of a dMod.frame
}
\examples{
\dontrun{
DL: TODO
}
}
