% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_cf}
\alias{theme_cf}
\title{Nice ggplot theme}
\usage{
theme_cf(base_size = 12, base_family = "", FLAGbold = TRUE)
}
\arguments{
\item{base_size, base_family}{see ?theme_bw}
}
\value{

}
\description{
Taken from dMod
}
