% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{todolist}
\alias{todolist}
\title{Update the TODO.md file in the base directory}
\usage{
todolist(wd = getwd(), FLAGpipes = TRUE, FLAGwrite_md = FALSE)
}
\arguments{
\item{FLAGpipes}{Boolean. search for pipe operators?}

\item{FLAGwrite_md}{}
}
\description{
Collect all TODO points in the scripts of a project and write them into TODO.md
}
\details{
todolist takes a directory and doesnt write the todolist to a file
}
