% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{toggle_off}
\alias{toggle_off}
\alias{toggle_on}
\alias{toggle}
\alias{toggle_section}
\alias{toggle_subsection}
\alias{toggle_subsubsection}
\alias{transform_subsection}
\alias{initiate_or_delete_subsection}
\alias{renumber_sections}
\title{Toggle sections}
\usage{
toggle_off(line, text, editor)

toggle_on(line, text, editor)

toggle(dashes = "----")

toggle_section()

toggle_subsection()

toggle_subsubsection()

transform_subsection(line, text, editor)

initiate_or_delete_subsection(line, text, editor)

renumber_sections(FLAGfunctionAsSection = TRUE)
}
\arguments{
\item{FLAGfunctionAsSection}{Start subsections within functions with 1. Switch off if the function is just a small part of a bigger logic}
}
\value{
called for side-effect
}
\description{
In the document outline, toggle a section on or off by adding or removing a #
}
\author{
Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)

Daniel Lill (daniel.lill@intiquan.com)
}
