% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{toggle_blabla}
\alias{toggle_blabla}
\title{Toggle BLABLA so the documentwalker can extract functions}
\usage{
toggle_blabla()
}
\value{
NULL: Modifies document
}
\description{
Toggle BLABLA so the documentwalker can extract functions
}
\details{
rstudioapi getSourceEditorContext documentSave setDocumentContents
}
