% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{toggle_mclapply_off}
\alias{toggle_mclapply_off}
\title{Title}
\usage{
toggle_mclapply_off(textline)
}
\arguments{
\item{textline}{}
}
\description{
Title
}
\examples{
textline <- "fuck <- mclapply(X = 1:n, mc.cores = ncores, FUN = function(x) {"
toggle_mclapply_off(textline)
}
