% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tpaste0}
\alias{tpaste0}
\title{Append a time-stamp before a string}
\usage{
tpaste0(...)
}
\arguments{
\item{mystring}{}
}
\description{
Append a time-stamp before a string
}
\examples{
tpaste0("workspace.rda")
}
