% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{trustAna_getArg}
\alias{trustAna_getArg}
\title{Get the i'th row of an argpath or argtry as a named vector}
\usage{
trustAna_getArg(fit, i = 1, whichPath = "argpath")
}
\arguments{
\item{fit}{a trust()-fit with blather = T}

\item{i}{the iteration (the step taken)}

\item{whichPath}{argpath or argtry, is regexed. as long as the regex matches uniquely, you can use it, e.g. "p" for path, "tr" for try}
}
\value{
named numeric
}
\description{
Get the i'th row of an argpath or argtry as a named vector
}
