% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{trustBlatherResults}
\alias{trustBlatherResults}
\title{Get blather results in nice data.table}
\usage{
trustBlatherResults(fit)
}
\arguments{
\item{fit}{return by dMod::trust(...,blather=TRUE)}
}
\value{

}
\description{
Get blather results in nice data.table
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
