% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understand_envirs.R
\name{understand_envirs}
\alias{understand_envirs}
\title{Function to understand environments}
\usage{
understand_envirs()
}
\description{
Function to understand environments
}
\examples{
map(1, function(i) {
mnv <- new.env()
cat("current position")
print(parent.env(mnv))
bla <- understand_envirs()
bla()
})
}
