% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{unlink_dMod}
\alias{unlink_dMod}
\title{Unlink standard dMod-generated files}
\usage{
unlink_dMod(endings = NULL)
}
\arguments{
\item{endings}{additional endings}
}
\description{
Removes all \*.c and \*.o from your workdir
}
