% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{unnest_listcol_dt}
\alias{unnest_listcol_dt}
\title{Unnest list column}
\usage{
unnest_listcol_dt(dt, colname)
}
\arguments{
\item{dt}{data.table with a list.column which is a list of data.tables}

\item{colname}{name of column with list(data.tables...)}
}
\value{
data.table with the list column unnested
}
\description{
Unnest list column
}
\examples{
dt <- data.table(a = 1:2, b = list(data.table(listdt1 = 1:3, listdt2 = 2:4), data.table(listdt1 = 4:6, listdt2 = 8:10) ))
colname <- "b"
dt1 <- unnest_listcol_dt(dt, colname)
dt1
}
