% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unnest_named}
\alias{unnest_named}
\title{Unnesting functions for list-columns}
\usage{
unnest_named(df, key, value)
}
\arguments{
\item{df}{a tibble with list columns}

\item{key, value}{like in gather and spread}
}
\description{
The following functions serve two purposes
1. Keep other list-columns when unnesting one specific list-column
2. Keep names of list-columns when unnesting them
3. Make it possible to spread a list-column
}
\details{
However, this inroduces the need for a key-column.
}
