% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{update_hypothesis}
\alias{update_hypothesis}
\title{Update a hypothesis}
\usage{
update_hypothesis(dMod.frame, exclude = NULL, include = NULL, prefix = "",
  suffix = "")
}
\arguments{
\item{suffix}{}
}
\description{
Update a hypothesis
}
\examples{
remove("doedel")
dMod.frame <- tibble(hypothesis = "one", doedel = "blabla", wup = "wup")

hypothesis = "one"
doedel <- "yay"
update_hypothesis(dMod.frame)
}
