% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{update_todolist}
\alias{update_todolist}
\title{Update the TODO.md file in the base directory}
\usage{
update_todolist()
}
\description{
Collect all TODO points in the scripts of a project and write them into TODO.md
}
