% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{update_version}
\alias{update_version}
\title{Update versions}
\usage{
update_version(cOde = T, dMod = T, cf = T, MRAr = F)
}
\arguments{
\item{cOde, dMod, cf, MRAr}{Logicals which packages are updated}
}
\description{
Update versions
}
