% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{zip_dMod.frame}
\alias{zip_dMod.frame}
\title{Zip a dMod.frame and its DLLs}
\usage{
zip_dMod.frame(dMod.frame, zipfile = NULL)
}
\arguments{
\item{dMod.frame}{The dMod.frame you want to zip}

\item{zipfile}{If you want to add the files to an existing zipfile, specify the filepath here.}
}
\description{
This might be useful for collaboration
}
