% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{zwrite}
\alias{zwrite}
\alias{zread}
\title{fwrite and zip it}
\usage{
zwrite(x, file, ...)

zread(file, ...)
}
\arguments{
\item{x}{}

\item{file}{}

\item{...}{}
}
\value{

}
\description{
fwrite and zip it
}
