% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{are_names_of}
\alias{are_names_of}
\title{Pipe-friendly assigning of vectors when their names are known first}
\usage{
are_names_of(char_vec, value, ...)
}
\arguments{
\item{char_vec}{Character}

\item{value}{Thre possibilities: 1. A function which takes \code{n} as an argument such as \code{rnorm}.
2. A vector of length 1, then this value gets recycled.
3. a vector of length \code{length(char_vec)}}

\item{...}{Arguments ging to value if value is a function}
}
\description{
Pipe-friendly assigning of vectors when their names are known first
}
\examples{
letters \%>\% are_names_of(rnorm)
letters \%>\% are_names_of(1)
letters \%>\% are_names_of(1:26)
}
