% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{as.datalist.datalist}
\alias{as.datalist.datalist}
\title{Identity function for datalists}
\usage{
\method{as.datalist}{datalist}(x, ...)
}
\arguments{
\item{x}{.}

\item{...}{.}
}
\description{
Identity function for datalists
}
