% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{as.parvec.tbl_df}
\alias{as.parvec.tbl_df}
\title{as.parvec for dMod.frames}
\usage{
\method{as.parvec}{tbl_df}(x, hypothesis = 1, index = 1)
}
\arguments{
\item{x}{dMod.frame}

\item{hypothesis}{1}

\item{index}{going to as.parvec.parframe}
}
\description{
as.parvec for dMod.frames
}
