% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_organization.R
\name{assign_folders}
\alias{assign_folders}
\title{Create output folders and assign names to variables}
\usage{
assign_folders(
  additional = NULL,
  scriptname = gsub(".R$", "", basename(rstudioapi::getSourceEditorContext()$path))
)
}
\arguments{
\item{scriptname}{character}
}
\description{
Create output folders and assign names to variables
}
