% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-full_setup.R
\name{cf_install_rstudio}
\alias{cf_install_rstudio}
\title{Quick setup of RStudio on a new machine}
\usage{
cf_install_rstudio(FLAGoverwrite = FALSE, FLAGshortcuts = FALSE,
  IQdesktopVersion = c("local", "Ueli"))
}
\value{
called for side-effect
}
\description{
Quick setup of RStudio on a new machine
}
