% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_parameterHandling.R
\name{cf_make_condition_specific}
\alias{cf_make_condition_specific}
\title{Make a column of an est.grid condition specific and augment est.vec}
\usage{
cf_make_condition_specific(est.grid, parameters_est_df, parname,
  conditions, condition_column = "ID", FLAGdummifyOtherConds = FALSE,
  condition.grid = NULL)
}
\arguments{
\item{est.grid, parameters_est_df}{as usual}

\item{parname}{character(1L) denoting a column in the est.grid}

\item{conditions}{character vector of conditions}

\item{FLAGdummifyOtherConds}{replace the other parameters by "dummy". Dummy will have value c(dummy = 1)}
}
\description{
new parameters will be named parname_"condition_column"
}
