% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_indiv.R
\name{cf_make_pars}
\alias{cf_make_pars}
\title{Extract pars, fixed and parnames from grids}
\usage{
cf_make_pars(pars, fixed = NULL, est.grid, fixed.grid, ID)
}
\arguments{
\item{est.grid}{needs condition and ID}

\item{fixed.grid}{}

\item{ID}{}

\item{est.vec}{}
}
\value{

}
\description{
Extract pars, fixed and parnames from grids
}
