% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cf_outputFigure}
\alias{cf_outputFigure}
\title{Output figures}
\usage{
cf_outputFigure(
  pl,
  filename,
  scriptname = basename(dirname(filename)),
  width = 14,
  height = 10,
  scale = 0.6,
  units = c("in", "cm", "mm"),
  dpi = 300,
  limitsize = TRUE,
  FLAGFuture = TRUE,
  ...
)
}
\arguments{
\item{pl}{plot like ggplot}

\item{filename}{}

\item{scriptname}{}

\item{width}{}

\item{height}{}

\item{scale}{}

\item{...}{}

\item{FLAGaddScriptname}{}
}
\value{

}
\description{
current features
* Add scriptname to plot as small caption
}
\examples{
inst/examples/S101-plotting.R
}
