% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cf_outputFigure}
\alias{cf_outputFigure}
\title{Output figures}
\usage{
cf_outputFigure(
  pl,
  filename,
  scriptname = basename(dirname(filename)),
  width = 14,
  height = 10,
  scale = 0.6,
  units = c("in", "cm", "mm"),
  dpi = 300,
  limitsize = TRUE,
  FLAGFuture = TRUE,
  ...
)
}
\arguments{
\item{pl}{plot like ggplot}

\item{filename}{}

\item{scriptname}{}

\item{width}{}

\item{height}{}

\item{scale}{}

\item{...}{}

\item{FLAGaddScriptname}{}
}
\value{

}
\description{
current features
* Add scriptname to plot as small caption
}
\examples{
# -------------------------------------------------------------------------#
# 0 Header ----
# -------------------------------------------------------------------------#
#
# S101-plotting.R
#
# [PURPOSE]
# 
# Try and demonstrate plotting with cfoutputFigure
#
#
# [AUTHOR]
# Daniel Lill
#
# [Date]
# Mon Mar 29 10:24:05 2021
#
.tempdir = tempdir()
.currentdir <- getwd()
setwd(.tempdir)

# library(conveniencefunctions)
devtools::load_all("~/Promotion/Promotion/Projects/conveniencefunctions")
rm(list = ls(all.names = TRUE))

.outputFolder <- paste0("../04-Output/", "S101-plotting")
for(folder in c(.outputFolder)) 
if(!dir.exists(folder)) dir.create(folder, recursive = TRUE)

# -------------------------------------------------------------------------#
# 1 Paginate ----
# -------------------------------------------------------------------------#
# Large facetted plot
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_grid_paginate(color ~ clarity, ncol = 3, nrow = 3, page = 4)

# Basic stuff
n_pages(pl)
getPaginateInfo(pl)
plotlist <- cf_applyPaginate(pl)
length(plotlist)

# Asynchronous outputting
cf_outputFigure(pl, filename = "wup.png")
cf_outputFigure(pl, filename = "wupwup.pdf")

# Go to the files
try(system("nautilus .", wait = FALSE))

# Exit ----
setwd(.currentdir)
future::plan("sequential")
}
