% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cf_outputFigure}
\alias{cf_outputFigure}
\title{Output figures}
\usage{
cf_outputFigure(
  pl,
  filename = NULL,
  width = 16,
  height = 10,
  scale = 1,
  units = c("cm", "mm", "in")[1],
  dpi = 300,
  limitsize = TRUE,
  FLAGFuture = TRUE,
  device = NULL,
  heightrel = NULL,
  ...
)
}
\arguments{
\item{pl}{plot like ggplot}

\item{filename, width, height, scale, units, dpi, limitsize, device, ...}{see [ggplot2::ggsave()]}

\item{FLAGFuture}{Export asynchronously with the future-package}
}
\value{
nothing
}
\description{
Handle paginate automatically, save asynchronously with future
}
\details{
# importFrom ggplot2 plot_dev parse_dpi plot_dim
}
\examples{
.tempdir = tempdir()
.currentdir <- getwd()
setwd(.tempdir)

library(conveniencefunctions)

# -------------------------------------------------------------------------#
# 1 Paginate ----
# -------------------------------------------------------------------------#
# Large facetted plot
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_grid_paginate(color ~ clarity, ncol = 3, nrow = 3, page = 4)

# Basic stuff
n_pages(pl)
getPaginateInfo(pl)
plotlist <- cf_applyPaginate(pl)
length(plotlist)

# Asynchronous outputting
cf_outputFigure(pl, filename = "wup.png")
cf_outputFigure(pl, filename = "wupwup.pdf")

# Go to the files
try(system("nautilus .", wait = FALSE))

# Exit ----
setwd(.currentdir)
future::plan("sequential")
}
