% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_indiv.R
\name{cf_predict}
\alias{cf_predict}
\title{predict.prdfn with options bowser and verbose}
\usage{
cf_predict(object, times, pars, ..., data = NULL, FLAGverbose = FALSE,
  FLAGbrowser = TRUE)
}
\arguments{
\item{object}{}

\item{times}{}

\item{pars}{}

\item{...}{}

\item{data}{}

\item{FLAGverbose}{}

\item{FLAGbrowser}{}
}
\description{
predict.prdfn with options bowser and verbose
}
