% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod_indiv.R
\name{cf_predict}
\alias{cf_predict}
\title{predict.prdfn with options bowser and verbose}
\usage{
cf_predict(prd, times, pars, keep_names = NULL, FLAGverbose = FALSE,
  FLAGverbose2 = FALSE, FLAGbrowser = FALSE, ...)
}
\arguments{
\item{times}{}

\item{pars}{}

\item{FLAGverbose}{}

\item{FLAGbrowser}{}

\item{...}{}

\item{object}{}

\item{data}{}
}
\description{
predict.prdfn with options bowser and verbose
}
