% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cf_profile_plotPathsAffected}
\alias{cf_profile_plotPathsAffected}
\title{Title}
\usage{
cf_profile_plotPathsAffected(profiles, page = 1, tol = 1e-05)
}
\arguments{
\item{profiles}{profiles to plotPaths}

\item{page}{pagination page}

\item{tol}{}
}
\value{
paginated ggplot
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
