% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRtools.R
\name{cf_script_templateVersions}
\alias{cf_script_templateVersions}
\title{Display template versions}
\usage{
cf_script_templateVersions(
  NfromLast = 3,
  path = ".",
  FLAGshortenIdentical = TRUE
)
}
\arguments{
\item{NfromLast}{Number of latest template versions to display}

\item{path}{as in list.files}
}
\value{

}
\description{
Display template versions
}
\examples{
path <- "/home/daniel/Promotion/Promotion/Projects/LiSyM/TGFb/Work/01-DynamicModeling/02-Scripts"
NfromLast = 3
FLAGshortenIdentical = TRUE
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
