% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{checkin_hypothesis}
\alias{checkin_hypothesis}
\title{Get elements from the .GlobalEnv and turn it into a row of the dMod.frame}
\usage{
checkin_hypothesis(dMod.frame, exclude = NULL, include = NULL,
  prefix = "", suffix = "")
}
\arguments{
\item{dMod.frame}{the dMod.frame}

\item{exclude}{not yet implemented}

\item{include}{not yet implemented}

\item{prefix}{prefix}

\item{suffix}{suffix

the dMod.frame augmented by the new row}
}
\description{
I don't really use this function, so it might not be worth it to keep working on it.
}
