% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{checkout_hypothesis}
\alias{checkout_hypothesis}
\title{Load one row of a dMod.frame into the .GlobalEnv}
\usage{
checkout_hypothesis(dMod.frame, hypothesis, prefix = "")
}
\arguments{
\item{hypothesis}{character or numeric. specifying the name  or the index of the hypothesis}
}
\description{
Load one row of a dMod.frame into the .GlobalEnv
}
\examples{
a <- 3
myfun <- function(x) {a * x^2}
testframe <- tibble(hypothesis = c("apap", "penner"),
                    schwurbel = list(qplot(1:10,1:10), "Penis"),
                    wupwup = list("moep", faithful),
                    myfun = list(function(x) {a * x^2}, myfun),
                    a = c(1:2))

checkout_hypothesis(testframe, "penner")
myfun
schwurbel
wupwup
hypothesis
}
