% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{construct_parframe}
\alias{construct_parframe}
\title{Helper function to reproducibly construct parframes}
\usage{
construct_parframe(pars, n = 20, seed = 12345, samplefun = rnorm, ...)
}
\arguments{
\item{pars}{Named vector. Values don't play a role, only names}

\item{n}{Integer how many lines should the parframe have}

\item{seed}{Seed for the random number generator}

\item{samplefun}{random number generator: \code{\link{rnorm}}, \code{\link{runif}}, etc...}

\item{...}{arguments going to samplefun}
}
\value{
parframe (without metanames)
}
\description{
Helper function to reproducibly construct parframes
}
\examples{
construct_parframe(c(a = 0, b = 100000), 5)
construct_parframe(c(a = 0, b = 100000), 5)
}
