% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{describe_plotValue}
\alias{describe_plotValue}
\title{Describe plot}
\usage{
describe_plotValue(dMod.frame = NULL)
}
\arguments{
\item{dMod.frame}{}
}
\description{
Here are some templates to motivate you to describe the plots.
}
