% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{describe_plotValue}
\alias{describe_plotValue}
\alias{describe_plotCombined}
\alias{describe_plotPars}
\alias{describe_plotPars}
\alias{describe_plotProfile}
\title{Insert description template}
\usage{
describe_plotValue(dMod.frame = NULL)

describe_plotCombined(dMod.frame = NULL)

describe_plotPars(dMod.frame = NULL)

describe_plotPars(dMod.frame = NULL)

describe_plotProfile(dMod.frame = NULL)
}
\arguments{
\item{dMod.frame}{}
}
\description{
Describe results in a Rmd-file with these templates.
}
