% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{getConditions.tbl_df}
\alias{getConditions.tbl_df}
\title{Try getting conditions}
\usage{
\method{getConditions}{tbl_df}(x, hypothesis = 1)
}
\arguments{
\item{x}{dMod frame}

\item{hypothesis}{1}
}
\description{
first from obj, then from data, then from p
}
