% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getReactionInfo}
\alias{getReactionInfo}
\title{Title}
\usage{
getReactionInfo(equationList, parInfo = getParInfo(equationList))
}
\arguments{
\item{el}{}

\item{p}{}
}
\value{
list of reactions
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
