% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_imports.R
\docType{import}
\name{ggly}
\alias{ggly}
\alias{reexports}
\alias{ggplotly}
\alias{getConditions}
\title{Reexports}
\usage{
ggly(p = ggplot2::last_plot(), width = NULL, height = NULL,
  tooltip = "all", dynamicTicks = FALSE, layerData = 1,
  originalData = TRUE, source = "A", ...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dMod}{\code{\link[dMod]{getConditions}}}

  \item{plotly}{\code{\link[plotly]{ggplotly}}}
}}

