% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyversetools.R
\name{ggshave}
\alias{ggshave}
\title{ggsave() with automatically saving metainformation of how the plot was produced}
\usage{
ggshave(filename, plot = last_plot(), device = NULL, path = NULL,
  scale = 1, width = NA, height = NA, units = c("in", "cm", "mm"),
  dpi = 300, limitsize = TRUE, ..., check_overwrite = T,
  commit_script = F, commit_plot = F)
}
\arguments{
\item{check_overwrite}{Asks before overwriting a plot}

\item{commit_script}{Lgl o character. Commit the script which generated th plot. Can be logical or commit message.}

\item{commit_plot}{Lgl o character. Commit the plot in its git repository. Can be logical or commit message.}

\item{all}{taken from ggsave()}
}
\description{
ggsave() with automatically saving metainformation of how the plot was produced
}
