% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets2.R
\name{insert_header}
\alias{insert_header}
\alias{insert_exit}
\title{Print a header to your console}
\usage{
insert_header(insert_in_script = T)

insert_exit(insert_in_script = T)
}
\arguments{
\item{insert_in_script}{boolean}
}
\description{
Print a header to your console
}
