% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_runbg}
\alias{insert_runbg}
\title{Insert a runbg job}
\usage{
insert_runbg(job_name = "myrunbg_job", job_type = NULL,
  dMod.frame = NULL, tol = 1, nsteps = 3, distribute = NULL)
}
\arguments{
\item{job_name}{character with "_job" at the end}

\item{job_type}{character, options: fit, profile, profile_steps}

\item{dMod.frame}{character "dMod.frame"}

\item{tol}{1 used for profile_steps}

\item{nsteps}{3}

\item{distribute}{vec of names of ssh machines c("knecht1", "knecht2")}
}
\description{
Insert a runbg job
}
\examples{
\dontrun{
insert_runbg()

insert_runbg(job_name = "myrunbg_job", job_type = "fit", dMod.frame = "myframe")
insert_runbg(job_name = "myrunbg_job", job_type = "profile", dMod.frame = "myframe")
myframe <- readRDS("~/Promotion/Projects/methacetin_fitting/Fit_model_41/2018_05_24_15_27_fit_krumbiegel_results.rds") \%>\% do.call(rbind,.)
insert_runbg(job_name = "myrunbg_job", job_type = "profile_steps", dMod.frame = "myframe", distribute = paste0("knecht", 1:5))
}

}
