% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{install_snippets}
\alias{install_snippets}
\title{Install cf snippets with snippr}
\usage{
install_snippets(remove_old = FALSE)
}
\arguments{
\item{remove_old}{should old snippets be removed? CAREFUL: deletes ALL custom snippets!}
}
\description{
Install cf snippets with snippr
}
