% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypublish.R
\name{mypublish_eqnvec2wordtable}
\alias{mypublish_eqnvec2wordtable}
\alias{print2word_eqnvec}
\title{Print the odes of an eqnvec in a tex and a docx table}
\usage{
mypublish_eqnvec2wordtable(x, path = "odes")

print2word_eqnvec(x, path = "odes")
}
\arguments{
\item{x}{eqnvec}
}
\description{
printout of c(x = -k\*x) will look like d/dt\[x] = - k*[x]
}
