% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{next_file}
\alias{next_file}
\title{Sequential file numbering}
\usage{
next_file(addition = "", base_name = "", fileext = "png",
  filepath = ".", purge = FALSE)
}
\arguments{
\item{basename, fileext, filepath, addition}{strings that are concatenated like this: paste0(filepath, basename, NUMBER, addition, fileext)}
}
\value{
character
}
\description{
Create filenames myfile001.png myfile002.png ... automatically
}
\examples{
nextfile()
}
\author{
dww from stackoverflow https://stackoverflow.com/questions/54752246/automatic-file-numbering-in-ggsave-as-in-png
}
