% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getMeasurementParsMapping}
\alias{petab_getMeasurementParsMapping}
\title{Parse the parameters columns in measurementData}
\usage{
petab_getMeasurementParsMapping(
  measurementData,
  column = c("observableParameters", "noiseParameters")[1]
)
}
\arguments{
\item{measurementData}{petab$measurementData. Uses only columns observableId,simulationConditionId and "column"}

\item{column}{character(1L). one of c("observableParameters", "noiseParameters")}
}
\value{
data.table(condition, Innerpars... = Outerpars...) of local parameters.
Can be added to a gridlist with \code{\link[dMod:indiv_addLocalParsToGridlist]{dMod::indiv_addLocalParsToGridlist()}}
}
\description{
Parse the parameters columns in measurementData
}
\examples{
measurementData <- data.table(
observableId = c("obsE"),
simulationConditionId = c("C1", "C2"),
observableParameters = c("scale;offset", "1;offset"),
noiseParameters = "sigmaAbs"
)

petab_getMeasurementParsMapping(measurementData, "noiseParameters")
petab_getMeasurementParsMapping(measurementData, "observableParameters")
# suggested use: indiv_addLocalParsToGridlist ... 
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
