% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{plotCombined.dMod.frame}
\alias{plotCombined.dMod.frame}
\title{Ideas for dMod}
\usage{
\method{plotCombined}{dMod.frame}(dMod.frame, hypothesis = 1, index = 1,
  ...)
}
\arguments{
\item{...}{}
}
\description{
in runbg, when starting eg 20 fits per core and the slowly the load decreases, because there are some fits that take ages, kill these few last fits
}
\details{
in normL2, when passing an errmodel with conditions, evalute only those conditions with errmodel and the rest without errmodel

plotting_functions: split them into a two-step pipe ("prepare_plotCombined" %>% "render_plotcombined") so that it can be more easily customizable?

str_detect_any

look at reticulate package
First way to OO plotting
}
