% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{plotCombined}
\alias{plotCombined}
\alias{plotCombined.tbl_df}
\alias{plotCombined.prdlist}
\title{Plot a list of model predictions and a list of data points in a combined plot}
\usage{
plotCombined(x, ...)

\method{plotCombined}{tbl_df}(x, hypothesis = 1, index = 1, ..., plotErrorBands = F)

\method{plotCombined}{prdlist}(
  x,
  data = NULL,
  ...,
  scales = "free",
  facet = "wrap",
  transform = NULL,
  aesthetics = NULL
)
}
\arguments{
\item{x}{prediction or dMod.frame}

\item{...}{to plotCombined.prdlist}

\item{hypothesis}{numeric length 1}

\item{index}{numeric refers to the fitrank of the fit stored in parframes of the dMod.frame}

\item{plotErrorBands}{TRUE: plot the errormodel as error bands?}

\item{data}{if x is prdlist, add datalist}

\item{scales}{to facet_wrap or grid}

\item{facet}{"wrap" or "grid"}

\item{transform}{coordinate transform}

\item{aesthetics}{named vector going to aes_() in the first call to ggplot. Allows very complex plots}
}
\description{
Plot a list of model predictions and a list of data points in a combined plot
}
\details{
Works with multiple steps (index as a vector). I recommend then using it with \code{aesthetics = c(linetype= "step_")}
to access the step-information
}
