% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{cf_plotCombined.dMod.frame}
\alias{cf_plotCombined.dMod.frame}
\alias{cf_plotCombined.prdlist}
\title{bla}
\usage{
cf_plotCombined.dMod.frame(x, hypothesis = 1, index = 1, ...,
  plotErrorBands = F)

cf_plotCombined.prdlist(x, data = NULL, ..., scales = "free",
  facet = "wrap", transform = NULL, aesthetics = NULL)
}
\description{
bla

bla
}
\details{
Works with multiple steps (index as a vector). I recommend then using it with \code{aesthetics = c(linetype= "step_")}
to access the step-information
}
