% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{prepare_argpath}
\alias{prepare_argpath}
\alias{plot_argpath}
\title{Look at the argpath of a fit}
\usage{
prepare_argpath(fit, ncluster = 1)

plot_argpath(argpath)
}
\arguments{
\item{fit}{a fit returned from dMod::trust()}

\item{argpath}{result from \link{prepare_argpath}}
}
\value{
a data.frame for plotting

a plot
}
\description{
Requires the fit to be run with blather = T
}
