% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_mathematica.character}
\alias{print_mathematica.character}
\title{Print a character in mathematica format}
\usage{
print_mathematica.character(mycharacter)
}
\arguments{
\item{mycharacter}{}
}
\description{
Print a character in mathematica format
}
\examples{
 \dontrun{print_mathematica.character(paste0(letters, "_", letters))}
}
\seealso{
Other printtools: 
\code{\link{print_mathematica.eqnlist}()},
\code{\link{print_mathematica.eqnvec}()},
\code{\link{print_mathematica.matrix}()},
\code{\link{print_mathematica.name_equals_value}()},
\code{\link{print_r.eqnlist2addReaction_pipe}()},
\code{\link{print_r.named_vector}()}
}
\concept{printtools}
