% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{print_mathematica.eqnlist}
\alias{print_mathematica.eqnlist}
\title{Print ode vector in mathematica format}
\usage{
print_mathematica.eqnlist(myeqnlist)
}
\arguments{
\item{myeqnlist}{an eqnlist}
}
\description{
Print ode vector in mathematica format
}
\examples{
NULL \%>\%
  addReaction("A", "B", "k1*A", "conversion A to B") \%>\%
  addReaction("B", "", "k2*B*A", "dedradation of B induced by A") \%>\%
  print_mathematica.eqnlist
}
