% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_mathematica.eqnvec}
\alias{print_mathematica.eqnvec}
\title{Print ode vector in mathematica format}
\usage{
print_mathematica.eqnvec(myeqnvec)
}
\arguments{
\item{myeqnvec}{an eqnvec}
}
\description{
Print ode vector in mathematica format
}
\examples{
NULL \%>\%
  addReaction("A", "B", "k1*A", "conversion A to B") \%>\%
  addReaction("B", "", "k2*B*A", "dedradation of B induced by A") \%>\%
  as.eqnvec \%>\%
  print_mathematica.eqnlist
}
