% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_mathematica.matrix}
\alias{print_mathematica.matrix}
\title{Print mathematica}
\usage{
print_mathematica.matrix(mat, name_in_mathematica = "M")
}
\arguments{
\item{mat}{}
}
\description{
Print mathematica
}
\examples{
\dontrun{print_mathematica.matrix(diag(1:5))}
}
\seealso{
Other printtools: \code{\link{print_mathematica.character}},
  \code{\link{print_mathematica.eqnlist}},
  \code{\link{print_mathematica.eqnvec}},
  \code{\link{print_mathematica.name_equals_value}},
  \code{\link{print_r.eqnlist2addReaction_pipe}},
  \code{\link{print_r.named_vector}}
}
\concept{printtools}
