% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_mathematica.name_equals_value}
\alias{print_mathematica.name_equals_value}
\title{Print the values such that they are assignments in mathematica}
\usage{
print_mathematica.name_equals_value(vec)
}
\arguments{
\item{vec}{}
}
\description{
Print the values such that they are assignments in mathematica
}
\examples{
\dontrun{print_mathematica.name_equals_value(c(a = 1, b = 2))}
}
\seealso{
Other printtools: 
\code{\link{print_mathematica.character}()},
\code{\link{print_mathematica.eqnlist}()},
\code{\link{print_mathematica.eqnvec}()},
\code{\link{print_mathematica.matrix}()},
\code{\link{print_r.eqnlist2addReaction_pipe}()},
\code{\link{print_r.named_vector}()}
}
\concept{printtools}
