% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mathematica_or_r.R
\name{print_r.named_vector}
\alias{print_r.named_vector}
\title{nicely formatted dput for named vectors}
\usage{
print_r.named_vector(myvec, indices = T)
}
\arguments{
\item{myvec}{}
}
\description{
nicely formatted dput for named vectors
}
\examples{
letters[1:5] \%>\% are_names_of(0) \%>\% print_r.named_vector
}
\seealso{
Other printtools: \code{\link{print_mathematica.character}},
  \code{\link{print_mathematica.eqnlist}},
  \code{\link{print_mathematica.eqnvec}},
  \code{\link{print_mathematica.matrix}},
  \code{\link{print_mathematica.name_equals_value}},
  \code{\link{print_r.eqnlist2addReaction_pipe}}
}
