% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print_r.named_vector}
\alias{print_r.named_vector}
\title{nicely formatted dput for named vectors}
\usage{
print_r.named_vector(myvec, indices = T)
}
\arguments{
\item{myvec}{}
}
\description{
nicely formatted dput for named vectors
}
\examples{
letters[1:5] \%>\% are_names_of(0) \%>\% print_r.named_vector
}
