% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{readPetab}
\alias{readPetab}
\title{Read PEtab files}
\usage{
readPetab(filename, FLAGTestCase = FALSE)
}
\arguments{
\item{FLAGTestCase}{}

\item{modelname}{}

\item{path}{}
}
\value{

}
\description{
Read PEtab files
}
