% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplotly}
\alias{reexports}
\alias{set_names}
\alias{reexports}
\alias{set_colnames}
\alias{reexports}
\alias{set_rownames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{set_names}}, \code{\link[magrittr]{set_colnames}}, \code{\link[magrittr]{set_rownames}}}

  \item{plotly}{\code{\link[plotly]{ggplotly}}}
}}

