% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_exportEquationList}
\alias{sbml_exportEquationList}
\title{Title}
\usage{
sbml_exportEquationList(
  equationList,
  filename,
  modelname = "Model",
  unitInfo = getUnitInfo(),
  speciesInfo = getSpeciesInfo(equationList),
  parInfo = getParInfo(equationList),
  compartmentInfo = getCompartmentInfo(equationList)
)
}
\arguments{
\item{equationList}{}

\item{filename}{}

\item{modelname}{}

\item{unitInfo}{}

\item{speciesInfo}{}

\item{parInfo}{}

\item{compartmentInfo}{}
}
\value{

}
\description{
Title
}
\examples{
modelname = "Model"
unitInfo = getUnitInfo()
speciesInfo = getSpeciesInfo(equationList)
parInfo = getParInfo(equationList)
compartmentInfo = getCompartmentInfo(equationList)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
