% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{simulate_data}
\alias{simulate_data}
\title{Simluate data with a dMod.frame}
\usage{
simulate_data(
  model,
  hypothesis = 1,
  output = c("dMod.frame", "data.frame", "datalist"),
  timesD = 0:10,
  s0 = 0.1,
  srel = 0.1,
  observables = getObservables(model, hypothesis),
  seed_pars = 1,
  seed_data = 1
)
}
\arguments{
\item{model}{dMod.frame, preferably with columns pars and covtable}

\item{hypothesis}{1}

\item{output}{output type. if dMod.frame, the current "data" in the given hypothesis will be overwritten}

\item{timesD, s0, srel, observables}{things to create the data template}

\item{seed}{}
}
\value{
either dMod.frame with data in it and column "truth" or data.frame
}
\description{
Simluate data with a dMod.frame
}
