% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{split_chunks}
\alias{split_chunks}
\title{Addin to insert tee-operator}
\usage{
split_chunks()
}
\description{
Addin to insert tee-operator
}
\seealso{
Other insertfunctions: \code{\link{insert_tee_operator}},
  \code{\link{insert_tee_print}},
  \code{\link{insert_unlink}},
  \code{\link{print_r.eqnlist2addReaction_pipe}}
}
\concept{insertfunctions}
