% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str0}
\alias{str0}
\alias{str1}
\alias{str2}
\alias{head1}
\alias{head3}
\title{str with max.level = 0}
\usage{
str0(...)

str1(...)

str2(...)

head1(...)

head3(...)
}
\arguments{
\item{...}{an object}
}
\description{
Saves you typoing when analysing objects
}
