% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr.R
\name{str_subset_keep_names}
\alias{str_subset_keep_names}
\title{str_subset that preserves names}
\usage{
str_subset_keep_names(vec, pattern)
}
\arguments{
\item{vec}{named vector}

\item{pattern}{regex as in str_subset}
}
\description{
str_subset that preserves names
}
\examples{
myvec <- c("a", "b", "ab", "ba") \%>\% are_names_of(.,.)
myvec \%>\% str_subset_keep_names("^a")
}
