% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todolist.R
\name{todolist}
\alias{todolist}
\title{Todolist}
\usage{
todolist(path = ".", FLAGrecursive = FALSE, FLAGremoveDone = TRUE)
}
\arguments{
\item{path}{path to search through scripts}

\item{FLAGrecursive}{search subfolders?}

\item{FLAGremoveDone}{remove done bullet points}
}
\value{
markdown table printed to console
}
\description{
Todolist
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{}
