% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{todolist}
\alias{todolist}
\title{Print a todolist to your console or to a file}
\usage{
todolist(path = getwd(), FLAGpipes = FALSE, filename = NULL)
}
\arguments{
\item{path}{path to directory to grep in}

\item{FLAGpipes}{Boolean. search for pipe operators?}

\item{filename}{If provided, write the todolist to this file (recommended ending: *.md)}
}
\description{
Collect all TODO points in the scripts of a project and write them into TODO.md
}
