% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_subsections.R
\name{toggle_off}
\alias{toggle_off}
\alias{toggle_on}
\alias{toggle}
\alias{toggle_section}
\alias{toggle_subsection}
\alias{toggle_subsubsection}
\title{Toggle sections}
\usage{
toggle_off(line, text, editor)

toggle_on(line, text, editor)

toggle(dashes = "----")

toggle_section()

toggle_subsection()

toggle_subsubsection()
}
\value{
called for side-effect
}
\description{
In the document outline, toggle a section on or off by adding or removing a #
}
