% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-toggle_subsections.R
\name{toggle_mclapply_on}
\alias{toggle_mclapply_on}
\title{Title}
\usage{
toggle_mclapply_on(textline)
}
\arguments{
\item{textline}{}
}
\value{

}
\description{
Title
}
\examples{
textline <- "fuck <- lapply(1:n, function(x) {"
toggle_mclapply_on(textline)
textline <- "fuck <- lapply(motherfuck[1:3], function(x) bla::fn(x))"
toggle_mclapply_on(textline)

}
