% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModframetools.R
\name{update_hypothesis}
\alias{update_hypothesis}
\title{Update a hypothesis}
\usage{
update_hypothesis(dMod.frame, exclude = NULL, include = NULL,
  prefix = "", suffix = "")
}
\arguments{
\item{dMod.frame}{bla}

\item{exclude}{bla}

\item{include}{bla}

\item{prefix}{bla}

\item{suffix}{bla}
}
\description{
Update a hypothesis
}
