% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{update_todolist}
\alias{update_todolist}
\alias{todolist}
\title{Update the TODO.md file in the base directory}
\usage{
update_todolist()

todolist(wd = getwd())
}
\description{
Collect all TODO points in the scripts of a project and write them into TODO.md
}
\details{
todolist takes a directory and doesnt write the todolist to a file
}
