% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModAugmentation.R
\name{wait_for_runbg}
\alias{wait_for_runbg}
\title{Wait for runbg}
\usage{
wait_for_runbg(job, delta_t = 5)
}
\arguments{
\item{delta_t}{}
}
\description{
It's better to use \code{runbg(..., wait = T)}, if you want to wait for the results anyway
}
\details{
But this was nice for practicing enquo()
}
